% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vis3D.R
\name{add_opacity}
\alias{add_opacity}
\title{Add opacity to a mesh}
\usage{
add_opacity(mesh, a_map, a_range = c(0.5, 1), a_default = 1)
}
\arguments{
\item{mesh}{The brain mesh of interest}

\item{a_map}{The vector of values to be used to set alpha (opacity)}

\item{a_range}{The range of alpha values to be used in the image (after rescaling),
must be between 0 and 1.}

\item{a_default}{the default alpha value for missing values in a_map}
}
\value{
The original mesh with the alpha levels set
}
\description{
Set the opacity for a brain mesh from a vector of values that correspond
to the vertices in the mesh.
}

