% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_hierarchical_anatomy.R
\name{addVolumesToHierarchy}
\alias{addVolumesToHierarchy}
\title{Add volumes to an anatomical hierarchy}
\usage{
addVolumesToHierarchy(hdefs, volumes)
}
\arguments{
\item{hdefs}{The anatomical hierarchy}

\item{volumes}{The matrix of volumes}
}
\value{
The input anatomical hierarchy with volumes added
}
\description{
Takes a tree of anatomical hierarchies and adds volumes from
\link{anatGetAll}
}
\details{
Each node in the tree has two new attributes: 
\itemize{
\item{volumes -}{ The matrix of volumes}
\item{meanVolume -}{ The mean of the volumes of that node}
}

Currently propagates volumes up the tree by summing. Future versions
  will add propagating through weighted means or other functions
}
\examples{
\dontrun{
abijson <- "allen.json"
defs <- "Dorr_2008_Steadman_2013_Ullmann_2013_mapping_of_labels.csv"
hdefs <- makeMICeDefsHierachical(defs, abijson)
allvols <- anatGetAll(gf$filenames)
hanat <- addVolumesToHierarchy(hdefs, allvols)
}
}
