% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_model_selection.R
\name{AICc}
\alias{AICc}
\title{Compute the Corrected AIC for an object}
\usage{
AICc(object, ...)
}
\arguments{
\item{object}{An object with an AICc method}

\item{...}{additional parameters for methods}
}
\description{
Corrected AIC for finite sample sizes. Generally recommended
over AIC as it asymptotically approaches AIC as n->infinity.
}
\details{
For use with \link{mincLm} objects AICc produces a vector of AICc values, one per voxel.
}

