% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_anatomy.R
\name{anatCreateVolume}
\alias{anatCreateVolume}
\title{Create a Statistic Volume}
\usage{
anatCreateVolume(anat, filename, column = 1)
}
\arguments{
\item{anat}{A minc volume object with a \code{labels} attribute}

\item{filename}{the filename for the new minc volume}

\item{column}{which column of the \link{anatLm} results to use}
}
\value{
Invisibly returns the created volume
}
\description{
Convert an anatomy volume into a statistic volume
with the results of an \link{anatLm}
}
