% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_FDR.R
\name{anatFDR}
\alias{anatFDR}
\alias{anatFDR.anatLmer}
\alias{anatFDR.anatModel}
\title{Anatomy False Discovery Rates}
\usage{
anatFDR(buffer, ...)

\method{anatFDR}{anatModel}(buffer, ...)

\method{anatFDR}{anatLmer}(buffer, ...)
}
\arguments{
\item{buffer}{the result of an \link{anatLm} type call}

\item{...}{additional parameters to \code{mincLmer} like method}
}
\value{
A object of type \code{mincQvals} with the same number of columns
as the input. Each column now contains the qvalues for each structure. Areas
outside the mask (if a mask was specified) will be represented by a
value of 1. The result also has an attribute called "thresholds"
which contains the 1, 5, 10, 15, and 20 percent false discovery rate
}
\description{
Anatomy False Discovery Rates
}
\section{Methods (by class)}{
\itemize{
\item \code{anatModel}: anatModel

\item \code{anatLmer}: anatLmerModel
}}

