% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_anatomy.R
\name{anatGetAll2}
\alias{anatGetAll2}
\title{Faster AnatGet}
\usage{
anatGetAll2(filenames, atlas, method = c("jacobians", "labels", "sums",
  "means"), side = c("both", "left", "right"), strict = TRUE)
}
\arguments{
\item{filenames}{... asdf}

\item{atlas}{... asdf}

\item{method}{... asdf}

\item{strict}{... asdf}

\item{mask}{... asdf}
}
\value{
something
}
\description{
doco
}

