% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_anatomy.R
\name{anatLmer}
\alias{anatLmer}
\title{Anatomy Linear Mixed Effects Model}
\usage{
anatLmer(formula, data, anat, REML = TRUE, control = lmerControl(),
  verbose = FALSE, start = NULL)
}
\arguments{
\item{formula}{the model formula}

\item{data}{the predictor variables}

\item{anat}{a subject by label matrix of anatomical 
summaries typically produced by \link{anatGetAll}}

\item{REML}{whether or not to use restricted expectation maximum likelihood
when fitting with \link[lme4]{lmer}}

\item{control}{configuration list for \link[lme4]{lmer}, typically generated
by \link[lme4]{lmerControl}.}

\item{verbose}{increase verbosity}

\item{start}{starting values for the optimizer in \link[lme4]{lmer}}
}
\value{
an \code{anatModel} object of statistical results
}
\description{
Apply a linear mixed effects model to the results of
\link{anatGetAll}.
}

