% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_anatomy.R
\name{anatLmerEstimateDF}
\alias{anatLmerEstimateDF}
\title{Estimate Degrees of freedom for an anatLmer model}
\usage{
anatLmerEstimateDF(model)
}
\arguments{
\item{model}{an \code{anatLmerModel}}
}
\value{
the same model, now with degrees of freedom set
}
\description{
Estimate the degrees of freedom for an \code{anatLmerModel} object
produced by \link{anatLmer}. See \link{anatLmer} for more details
on degrees of freedom estimation for linear mixed effects models
}

