% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_anatomy.R
\name{anatSummarize}
\alias{anatSummarize}
\title{Summarize anatGetAll results by hierarchy}
\usage{
anatSummarize(anat, summarize_by = "hierarchy",
  defs = getOption("RMINC_LABEL_DEFINITIONS"), discard_missing = FALSE)
}
\arguments{
\item{anat}{The results of a call to \link{anatGetAll}}

\item{summarize_by}{either a data frame with grouping information or a path to label definitions. 
If a data frame is passed it must have two columns: "label" containing the structure
name (see \code{colnames(anat)} to check the relevant labels) and "grouping" containg the name 
of the group each structure belongs to.}

\item{discard_missing}{logical controlling how to handle structures with no ("") group information.
If TRUE filter these structures, if FALSE give each structure a group label to match their structure
name.}
}
\value{
A matrix with columns representing groups from the hierarchy and rows representing individuals
with values equal to the sum of the individual members of each group.
}
\description{
Take the results of \link{anatGetAll} and summarize a grouping label, typically the structures
group in the anatomical hierarchy.
}
