% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civet.R
\name{civet.CreateBrainViewROI}
\alias{civet.CreateBrainViewROI}
\title{civet.CreateBrainViewROI}
\usage{
civet.CreateBrainViewROI(atlasFile, atlasVertices, region,
  civetVersion = "1.1.12")
}
\arguments{
\item{atlasFile}{File with Atlas label (string-label(int)) mappings}

\item{atlasVertices}{File with vertex label(int) mappings}

\item{region}{Region to plot -> Must match name in atlas file exactly}

\item{civetVersion}{Version of CIVET}
}
\description{
Create a brainview file of a specific ROI
}
\details{
Create a .txt file that can be overlaid unto the model template (usually in the CIVET models directory)
Currently only CIVET version 1.1.12 is supported.
}
\examples{
\dontrun{
getRMINCTestData() 
civet.CreateBrainViewROI("/tmp/rminctestdata/AAL.csv",
                         "/tmp/rminctestdata/AAL_atlas_left.txt",
                         "Left Insula")
q()
#(The .txt file is written in the working directory and can be viewed via the command)
#brain-view2 $CIVET_DIR/models/surf_reg_model_left.obj ./LeftInsula.txt
}
}
\seealso{
civet.CreateBrainViewFile
}
