% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civet.R
\name{civet.flattenForDplyr}
\alias{civet.flattenForDplyr}
\title{Flatten CIVET results for dplyr}
\usage{
civet.flattenForDplyr(civetResults, columnsToKeep)
}
\arguments{
\item{civetResults}{A data.frame produced by \link{civet.readAllCivetFiles}}

\item{columnsToKeep}{vector of column names or indices of columns from the original
frame to copy to the normalized table. Columns not produced by \link{civet.readAllCivetFiles}
and not specified here are dropped. See details}
}
\value{
data.frame containing results of \link{civet.readAllCivetFiles}, all sub-data.frames
and matrices are expanded, non-standard characters in column names are replaced with underscores,
and a prefix denoting the origin sub-data is given.
}
\description{
Convert the data.frame/Matrix/list fusion object produced by civet.readAllCivetFiles
to a data.frame usable with dplyr etc.
}
\details{
The columnsToKeep vector needs to include the subject identifier or it will be
dropped. Ideally other columns in the vector should be proper vector/list columns compatible
with dplyr
}
