% Generated by roxygen2 (4.0.1): do not edit by hand
\name{civet.getAllFilenames}
\alias{civet.getAllFilenames}
\alias{civet.organizeCivetDatFilesMidWhiteGrey}
\title{civet.getAllFilenames}
\usage{
civet.getAllFilenames(gf, idvar, prefix, basedir, append = TRUE, civetVersion = "1.1.9")

civet.getAllFilenames(gf, idvar, prefix, basedir, append = TRUE, civetVersion = "1.1.9")
}
\arguments{
\item{gf}{Data Frame with subject information}

\item{idvar}{column name in gf with subject IDs}

\item{prefix}{Prefix specified when CIVET was run}

\item{basedir}{directory where all CIVET output was stored}

\item{append}{Whether to append the results to the input gf}

\item{civetVersion}{Version of CIVET}

\item{gf}{Data Frame with subject information}

\item{idvar}{column name in gf with subject IDs}

\item{prefix}{Prefix specified when CIVET was run}

\item{basedir}{directory where all CIVET output was stored}

\item{append}{Whether to append the results to the input gf}

\item{civetVersion}{Version of CIVET}
}
\value{
gf is returned with CIVET filenames

gf is returned with CIVET filenames

mincLm Returns a vector of mincSingleDim class
}
\description{
Generates list of filenames output by CIVET

Generates list of filenames output by CIVET
}
\details{
Prior to running, read.csv  may be called to generate the input argument gf.
The results will be stored under the column name CIVETFILES either in the input gf (if append = TRUE) or in a new gf.
Currently only CIVET versions 1.1.9 and 1.1.12 are supported.

Prior to running, read.csv  may be called to generate the input argument gf.
The results will be stored under the column name CIVETFILES either in the input gf (if append = TRUE) or in a new gf.
Currently only CIVET versions 1.1.9 and 1.1.12 are supported.
}
\examples{
gf = read.csv("~/SubjectTable.csv")
civet.getAllFilenames(gf,"ID","ABC123","~/CIVET","TRUE","1.1.12")
gf = civet.readAllCivetFiles("~/Atlases/AAL/AAL.csv",gf)
gf = read.csv("~/SubjectTable.csv")
civet.getAllFilenames(gf,"ID","ABC123","~/CIVET","TRUE","1.1.12")
gf = civet.readAllCivetFiles("~/Atlases/AAL/AAL.csv",gf)
}
\seealso{
civet.readAllCivetFiles

civet.readAllCivetFiles
}

