% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civet.R
\name{civet.organizeCivetDatFilesAtlas}
\alias{civet.organizeCivetDatFilesAtlas}
\title{Organizes CIVET .dat files based on an Atlas}
\usage{
civet.organizeCivetDatFilesAtlas(atlasFile, dataFiles, civetVersion = "1.1.12")
}
\arguments{
\item{atlasFile}{Character path to a key to the atlas used when running civet.
the key should be a comma separated file with a header and the following form \cr
Column 1: Numeric label
Column 3: Corresponding structure}

\item{dataFiles}{character containing paths to .dat files of interest
typically generated with \link{civet.getAllFilenames}}

\item{civetVersion}{character code for the version of civet used to generate
the data files}
}
\description{
Uses an atlas to associate the measurement results in
the CIVET .dat output files with particular structures
}
