% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civet.R
\name{civet.readCBRAIN}
\alias{civet.readCBRAIN}
\title{Read a CBRAIN project in R}
\usage{
civet.readCBRAIN(path, prefix, subjects = NULL, atlas = "AAL",
  civetVersion = "2.1.0", readFiles = TRUE, readQC = TRUE,
  flatten = TRUE, QCDir = "QC")
}
\arguments{
\item{path}{Path to the civet project}

\item{prefix}{The prefix used to create the subject names}

\item{subjects}{A character vector specifying which subjects to read in. If not specified, All files within
\code{path} that don't match QC, scripts, stats, or analysis will be read in.}

\item{atlas}{Either AAL (default), DKT, or a path to the specific atlas used}

\item{civetVersion}{the version of CIVET used, either 2.1.0 (default) or 2.0.0}

\item{readFiles}{logical whether or not to read the files into R or just generate the file names}

\item{readQC}{logical whether to read and merge the QC results (must be used with \code{flatten})}

\item{flatten}{logical whether to convert the CIVET results into a \code{dplyr} compatible data frame or leave
it in the legacy format}

\item{QCDir}{The directory, or vector of directories of where to find QC tables}
}
\value{
A data.frame in the format of \link{civet.getAllFilenames} if \code{readFiles} is FALSE. A data.frame in
the format of \link{civet.readAllCivetFiles} if \code{readFiles} is TRUE and \code{flatten} is FALSE. And a data.frame in
the format of \link{civet.flattenForDplyr} if \code{readQC}, \code{readFiles}, and \code{flatten} are all TRUE (default)
}
\description{
Simplified file input for CBRAIN projects. Attempts to use as much information as possible from the directory structure
and internal CBRAIN config files to find and read the correct CIVET results for CIVET 2.0.0 and 2.1.0
}
\seealso{
\link{civet.getAllFilenames} \link{civet.readAllCivetFiles} \link{civet.flattenForDplyr}
}
