% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civet.R
\name{civet.readQC}
\alias{civet.readQC}
\title{Read CIVET QC data}
\usage{
civet.readQC(dir, civetVersion = "1.1.12")
}
\arguments{
\item{dir}{The CIVET output directory}

\item{civetVersion}{the version of CIVET used, currently only supports
1.1.12, 2.0.0, 2.1.0}
}
\value{
A table of QC results including whether or not the subjects passed
overall quality control. See \url{http://www.bic.mni.mcgill.ca/ServicesSoftware/QualityControlCIVET12}
for more details.
}
\description{
After running the CIVET quality control pipeline, import the results
}

