% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civet.R
\name{civet.vertexFilenames}
\alias{civet.vertexFilenames}
\title{Assemble vertex files for a CIVET run}
\usage{
civet.vertexFilenames(
  gf,
  idvar,
  prefix,
  basedir,
  append = TRUE,
  civetVersion = "1.1.9"
)
}
\arguments{
\item{gf}{Data Frame with subject information}

\item{idvar}{column name in gf with subject IDs}

\item{prefix}{Prefix specified when CIVET was run}

\item{basedir}{directory where all CIVET output was stored}

\item{append}{Whether to append the results to the input gf}

\item{civetVersion}{Version of CIVET}
}
\value{
A data.frame containing left and right thickness, area, and volume files.
}
\description{
Locate the vertex thickness, area, and volume files for a CIVET run
}
