% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civet.R
\name{civet.vertexTable}
\alias{civet.vertexTable}
\title{Create a table of vertex measures}
\usage{
civet.vertexTable(vertex_files)
}
\arguments{
\item{vertex_files}{The data frame of vertex file names
produced by \link{civet.vertexFilenames}}
}
\value{
A 6-element list of matrices:
\itemize{
\item{}
}
}
\description{
Read in the vertex data results of a CIVET run
}
