% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_on_manifolds.R
\name{clean_up_manifold_triangles}
\alias{clean_up_manifold_triangles}
\title{Clean up manifold mesh}
\usage{
clean_up_manifold_triangles(manifold, k = 1)
}
\arguments{
\item{manifold}{A list of length 2 with names 'vertex_matrix' and 'triangle_matrix'
like \code{bic_obj} object produced by \link{read_obj}}

\item{k}{degrees-of-seperation for defining neighbours. Default is 1 (i.e. adjacent vertices)}
}
\value{
A list of length 2 with names 'vertex_matrix' and 'triangle_matrix' 
similar to 'manifold' argument. Problematic triangles (if they exist) are removed from the mesh.
}
\description{
If mesh edge has 2+ attached triangles, keep the 2 triangles most well-connected triangles
}
\details{
increasing k drastically slows down computation
}
