% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vis3D.R
\name{colour_mesh}
\alias{colour_mesh}
\title{Colourize a mesh}
\usage{
colour_mesh(mesh, colour_map, colour_range = NULL, colour_default = "grey",
  symmetric = NULL, labels = FALSE, palette = heat.colors(255))
}
\arguments{
\item{mesh}{\link[rgl]{mesh3d} object ideally produced by \link{create_mesh}}

\item{colour_map}{either a vector with a label/measure/statistic for every vertex
or a character file path pointing to a file with such a vector in a rowwise format.}

\item{colour_range}{a two element numeric vector indicating the min and max values of 
allowable labels/measures/statistics to be includedon the surface}

\item{colour_default}{The colour given to vertices excluded by colour_range}

\item{symmetric}{Whether to have a positive and negative colour scale (not yet implemented)}

\item{labels}{Whether or not the colour_map is a set of discrete labels}

\item{palette}{A palette, AKA look-up-table, providing a linear colour scale for the colours in
\code{colour_map}}
}
\value{
an \code{obj_mesh} object descended from \link[rgl]{mesh3d}, with added colour information
and an additional \code{legend} element to be used in building a colour bar
}
\description{
Add colour information to your mesh, either from a vertex atlas like AAL
or from a statistic/measurement map like those produced by CIVET
}
