% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_model_selection.R
\name{compare_models}
\alias{compare_models}
\title{Compare a set of massively-univariate models}
\usage{
compare_models(object, ..., metric = AICc)
}
\arguments{
\item{object}{A model or a list of models, typically 
\link{mincLm}, \link{vertexLm}, or \link{anatLm} results.}

\item{...}{additional models}

\item{metric}{A function to apply to the models that extracts a result
for each independent sub-model. Typical choices are \link{AIC}, \link{AICc}, 
and \link{BIC}}
}
\value{
A sub-model x n models \code{model_comparison} matrix with the metric of interest.
}
\description{
For each response (voxel, vertex, or structure) compare
a set of linear model formulations with the criterion of your
choice (e.g. AIC, AICc, BIC).
}

