% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_graphs.R
\name{components_to_map}
\alias{components_to_map}
\title{Components to map}
\usage{
components_to_map(graph, components, type = c("size", "label"))
}
\arguments{
\item{graph}{the parent \code{igraph} generated by \code{obj_to_graph}
and used with \code{connected_components}}

\item{components}{the data.frame results of \code{connected_components}}

\item{type}{whether to return the size of the component or its label}
}
\value{
A vector corresponding to each vertex in \code{graph} with either
zero, or the size/label from the connected components calculation
}
\description{
Convert a \link{connected_components} data.frame into a vertex map
for visualizing the results.
}
