% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_parallel.R
\name{configureMincParallel}
\alias{configureMincParallel}
\title{Parallel Configuration}
\usage{
configureMincParallel(queue_type = c("custom", "sge", "pbs"),
  config_file = NULL)
}
\arguments{
\item{queue_type}{one of custom, sge, or pbs (torque)}

\item{config_file}{the configuration file to use if a custom \code{queue_type} is
selected.}
}
\value{
returns the configuration list invisibly
}
\description{
Set up configuration for running parallel jobs with RMINC
Use a standard configuration for SGE and Torque queuing systems
or setup a custom configuration. Essentially a wrapper to
\link{loadConfig} from BatchJobs, with two standard configurations
provided.
}
\details{
In order for parallelization to be as general as possible we provide means to setup the parallelization
backend (BatchJobs) for any type of queuing system. For details on how to configure BatchJobs see
\url{https://github.com/tudo-r/BatchJobs/wiki/Configuration}, or get inspiration from the SGE and PBS
template scripts provided for SGE and PBS are located in \code{system.file("parallel/", package = "RMINC")}.
Use of this function can be unnecessary, RMINC honours the hierarchy
of configuration files <BatchJobs>/.BatchJobs.R < ~/.BatchJobs.R < getwd()/.BatchJobs.R
you can edit one of these files to suit your use case. Alternatively, you can modify
your R \link{Startup} files with \code{
setHook(packageEvent("RMINC", "attach"),
function(...) 
  configureMincParallel("custom", "<your-config-file.R>"))
} to set your configuration whenever \code{RMINC} is loaded
}

