% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vis3D.R
\name{connected_components}
\alias{connected_components}
\title{Find connected components of a object graph}
\usage{
connected_components(obj_graph, data_map, data_range)
}
\arguments{
\item{obj_graph}{The \code{igraph} object generated by \link{obj_to_graph}}

\item{data_map}{A set of data corresponding to the vertices}

\item{data_range}{A two-component vector giving lower and upper bounds
for vertices to include when generating connected components}
}
\value{
a data frame of vertices after thresholding indicating the
vertex id, the cluster it's a member of, and the size of its cluster
}
\description{
Given vertex data and a threshold, find the connected
component of a graph representing a \code{bic_obj}
}

