% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_functions.R
\name{create_mesh}
\alias{create_mesh}
\title{Create surface mesh for a bic_obj}
\usage{
create_mesh(bic_obj, color = "grey", specular = "black",
  add_normals = TRUE, ...)
}
\arguments{
\item{bic_obj}{A bic_obj to be converted to a mesh}

\item{color}{a character indicating the colour of triangles in the mesh}

\item{specular}{a character indicating the colour of the specular the lighting produces
on the object, "black" prevents the objects from looking shiny}

\item{add_normals}{logical whether to add normals to the mesh to make the surface less jagged}

\item{...}{extra parameters passed to the material argument list of \link{tmesh3d} can include
additional \link{rgl.material} parameters to give to your mesh}
}
\value{
a \code{obj_mesh} descended from \link{mesh3d} object to be
plotted alone or subsequently colourized with \link{colour_mesh}
}
\description{
Turn the bic_obj object into an rgl tmesh ready for visualization
}

