% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{extract_column}
\alias{extract_column}
\title{Internal function to read in a table or csv file 
potentially with many columns 
and extract just the one we need}
\usage{
extract_column(filename, column = 1)
}
\arguments{
\item{filename}{path to the vertex data file}

\item{column}{-specify the column id (name or number) if input files have multiple columns}
}
\value{
a vector of vertex data corresponding to the file
}
\description{
Internal function to read in a table or csv file 
potentially with many columns 
and extract just the one we need
}
