% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{getRMINCTestData}
\alias{getRMINCTestData}
\title{Download Example Data}
\usage{
getRMINCTestData()
}
\description{
Download the example data needed to run our examples in your /tmp directory
The data can be downloaded manually from 
\url{https://wiki.mouseimaging.ca/download/attachments/1654/rminctestdata.tar.gz}
}

