% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_hierarchical_anatomy.R
\name{hanatAnova}
\alias{hanatAnova}
\title{ANOVA applied to every item in an anatomical tree}
\usage{
hanatAnova(formula, data, anatTree)
}
\arguments{
\item{formula}{The model formula}

\item{data}{The data frame containing the right side of the formula}

\item{anatTree}{The anatomical tree}
}
\value{
The anatomical tree with the model information added
}
\description{
See \link{anatAnova} for more detail
}
\details{
The volumes inside the anatomical hierarchy and the data must be of
  the same length and ordering.
}
\examples{
\dontrun{
vols <- addVolumesToHierarchy(hdefs, allvols)
hLm <- hanatAnova(~Sex, gf, vols)
}
}
