% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_hierarchical_anatomy.R
\name{hanatFDR}
\alias{hanatFDR}
\title{Compute the False Discovery Rate for an anatomical hierarchy}
\usage{
hanatFDR(buffer)
}
\arguments{
\item{buffer}{What to compute FDR on}
}
\value{
the anatomical hierarchy with FDR added (see details)
}
\description{
Compute the False Discovery Rate for an anatomical hierarchy
}
\details{
Uses the Benjamini & Yekutieli (2001) algorithm for computing FDR to
  account for dependence (which is obviously present in a hierarchy where 
  parent values are directly derived from their children).
  
  Stores the values using the same name as the incoming statistic but with a
  q prepended. I.e. tvalue.Sexmale becomes qtvalue.Sexmale. Note that the original, 
  input, is modified with these new additions as well.
}
\examples{
\dontrun{
vols <- addVolumesToHierarchy(hdefs, allvols)
hLm <- hanatLm(~Sex, gfBasic, vols)
hLm <- hanatFDR(hLm)
thresholds(hLm)
}
}
