% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_hierarchical_anatomy.R
\name{hanatLmer}
\alias{hanatLmer}
\title{linear mixed effects model applied to every item in an anatomical tree}
\usage{
hanatLmer(formula, data, anatTree, ...)
}
\arguments{
\item{formula}{The model formula}

\item{data}{The data frame containing the right side of the formula}

\item{anatTree}{The anatomical tree}

\item{...}{Extra options (REML, control of parallel execution, etc.) passed on to anatLmer}
}
\value{
The anatomical tree with the model information added
}
\description{
See \link{anatLmer} for more detail
}
\details{
The volumes inside the anatomical hierarchy and the data must be of
  the same length and ordering.
}
\examples{
\dontrun{
vols <- addVolumesToHierarchy(hdefs, allvols)
hLm <- hanatLmer(~Sex + (1|ID), gf, vols)
}
}
