% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_hierarchical_anatomy.R
\name{hanatLmerEstimateDF}
\alias{hanatLmerEstimateDF}
\title{Estimates degrees of freedom}
\usage{
hanatLmerEstimateDF(buffer, n = 50)
}
\arguments{
\item{buffer}{input hierarchical tree}

\item{n}{number of structures to use for DF estimation}
}
\value{
input tree with df added
}
\description{
see \link{anatLmerEstimateDF}
}
\details{
The volumes inside the anatomical hierarchy and the data must be of
  the same length and ordering.
}
\examples{
\dontrun{
vols <- addVolumesToHierarchy(hdefs, allvols)
hLm <- hanatLmer(~Sex + (1|ID), gf, vols)
hLm <- hanatLmerEstimateDF(hLm)
hLm <- hanatFDR(hLm)
thresholds(hLm)
}
}
