% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_hierarchical_anatomy.R
\name{hanatPlot}
\alias{hanatPlot}
\title{hanatPlot}
\usage{
hanatPlot(anatTree)
}
\arguments{
\item{anatTree}{the input anatomical tree}
}
\description{
Create a basic plot of the anatomy tree
}
