% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_hierarchical_anatomy.R
\name{hanatToVisGraph}
\alias{hanatToVisGraph}
\alias{hanatView}
\title{Turns an anatomical tree into a nodes and edges for visNetwork plotting}
\usage{
hanatToVisGraph(hanatTree, colourVariable = "color_hex_triplet",
  colourScale = colorRampPalette(c("red", "yellow"))(255),
  rColourScale = colorRampPalette(c("blue", "turquoise1"))(255), low = NULL,
  high = NULL, symmetric = F, transparent = "#FDFDFD",
  edgeColourFromABI = F)

hanatView(...)
}
\arguments{
\item{hanatTree}{The input anatomical tree}

\item{colourVariable}{String containing the variable to colour nodes with}

\item{colourScale}{Colour scale to use}

\item{rColourScale}{Reverse colour scale to use if symmetric colours enabled}

\item{low}{Lower cut-off for colour scale}

\item{high}{Upper cut-off for colour scale}

\item{symmetric}{Boolean for whether colour scale is symmetric}

\item{transparent}{Colour to use for transparent nodes}

\item{edgeColourFromABI}{Whether to set edge colours based on ABI atlas}
}
\value{
list with nodes and edges data frames
}
\description{
Turns an anatomical tree into a nodes and edges for visNetwork plotting
}
\section{Functions}{
\itemize{
\item \code{hanatView}: Directly view the graph
}}

