% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_hierarchical_anatomy.R
\name{hanatToVolume}
\alias{hanatToVolume}
\title{Converts a column from a hierarchical tree into a volume for viewing or saving}
\usage{
hanatToVolume(anatTree, labelVolume, column)
}
\arguments{
\item{anatTree}{The anatomical tree}

\item{labelVolume}{The volume containing the label definitions}

\item{column}{String indicating which column to turn into a volume}
}
\value{
The volume with the values from the anatomical tree
}
\description{
Converts a column from a hierarchical tree into a volume for viewing or saving
}
\examples{
\dontrun{
labelVol <- mincArray(mincGetVolume("some-labels.mnc"))
hLm <- hanatLm(~Sex, gfBasic, vols)
statsvol <- hanatToVolume(hLm, labelVol, "F.statistic")
mincPlotSliceSeries(anatVol, statsvol, anatLow = 700, anatHigh = 1400, 
  low=1, high=10, symmetric = F, begin=50, end=-50)
}
}
