% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{is.minc}
\alias{is.minc}
\title{Check for RMINC objects}
\usage{
is.minc(x)
}
\arguments{
\item{x}{the object of interest}
}
\value{
logical whether or not the object is an RMINC object
}
\description{
Checks if the current object is one produced by RMINC
(currently \code{mincList}, \code{mincSingleDim}, \code{mincMultiDim})
}
