% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_on_manifolds.R
\name{laplace_beltrami_operator}
\alias{laplace_beltrami_operator}
\title{Compute Laplace-Beltrami operator}
\usage{
laplace_beltrami_operator(
  manifold = NULL,
  vertex_matrix = NULL,
  triangle_matrix = NULL
)
}
\arguments{
\item{manifold}{A list of length 2 with names 'vertex_matrix' and 'triangle_matrix'
like \code{bic_obj} object produced by \link{read_obj}}

\item{vertex_matrix}{3-by-N matrix denoting the position of the N vertices 
defining the triangle-mesh manifold. Similar to argument \code{vertices} in
\link[rgl:mesh3d]{rgl::tmesh3d}. Required if \code{manifold} argument is not specified.}

\item{triangle_matrix}{3-by-M matrix denoting the position of the M triangles
defining the triangle-mesh manifold. Elements of each column are indices of vertices 
defining the triangle (i.e. indices are columns of the vertex_matrix). Similar to
argument \code{indices} in \link[rgl:mesh3d]{rgl::tmesh3d}. 
Required if \code{manifold} argument is not specified.}
}
\value{
sparse square matrix representing the discrete Laplace-Beltrami operator
for the manifold
}
\description{
Discrete Laplace-Beltrami operator associated with a triangle-mesh manifold
}
\details{
Either supply \code{manifold} argument OR \code{vertex_matrix} and \code{triangle_matrix}
arguments.
}
