% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_shiny.R
\name{launch_shinyRMINC}
\alias{launch_shinyRMINC}
\title{launch a shiny based inspector}
\usage{
launch_shinyRMINC(statsoutput, anatVol, volumes = NULL,
  keepBetas = FALSE, plotcolumns = NULL, modelfunc = NULL,
  singleStatType = NULL, fdr = NULL, anatLow = 700,
  anatHigh = 1400)
}
\arguments{
\item{statsoutput}{the output of mincLm, mincAnova, or mincLmer. Alternatively
a single statistic volume. Must be compatible with \link{mincArray}}

\item{anatVol}{the anatomical volume on which to display the statistics. Can
be passed a filename, a MINC volume (from mincGetVol), or a mincArray.}

\item{volumes}{a matrix or data frame of volumes for plotting}

\item{keepBetas}{whether to include the beta coefficients}

\item{plotcolumns}{extra data to be used for plotting}

\item{modelfunc}{optional modelling function}

\item{singleStatType}{if passing a single statistic volume to \code{statsoutput}
what statistic type is it. Stat-types "b", "t", and "tlmer" will get symmetric
colour scales. All others will get one sided colour scales.}

\item{fdr}{An optional mincFDR object for choosing thresholds.}

\item{anatLow}{The lower threshold value for displaying the underlying anatomy}

\item{anatHigh}{The upper threshold value for displaying the underlying anatomy}
}
\description{
This uses shiny to inspect the output of mincLm or mincAnova. Produces
various views, the ability to plot individual voxels, and get the output of
the stats model at that voxel.
}
\examples{
\dontrun{
vs <- mincLm(reljacobians02 ~ sex*treatment, subset(gfs, treatment != "None"))
anatVol <- mincArray(mincGetVolume("anatomyfile.mnc"))
launch_shinyRMINC(vs, anatVol, volumes=gfs$vols, 
                  plotcolumns=gfs[,c("sex", "Neonatal")], keepBetas=F)
}
}
