% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vis3D.R
\name{line_obj_to_voxel}
\alias{line_obj_to_voxel}
\title{Convert Lines to Voxel Coordinates}
\usage{
line_obj_to_voxel(line_obj, minc_file)
}
\arguments{
\item{line_obj}{The \code{bic_lines} object of interest}

\item{minc_file}{The reference file for computing voxel coordinates}
}
\value{
A \code{bic_lines} object in voxel coordinates
}
\description{
Convert a \code{bic_lines} object to world coordinates for plotting
}
