% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vis2D.R
\name{lut_to_palette}
\alias{lut_to_palette}
\title{A tool that returns a color function/palette from color lookup files}
\usage{
lut_to_palette(lookup_table = system.file("luts/spectral", package = "RMINC"),
  alpha = 1)
}
\arguments{
\item{lookup_table}{Either a path to the lookup table file, or the table itself}

\item{alpha}{A transparency value between 0 and 1 (inclusive)}
}
\value{
A function that takes in an integer specifying the number of colours required, and returns a vector of colours interpolated from the input colour lookup file
}
\description{
A tool that returns a color function/palette from color lookup files
}
\note{
For now, the first column of the lookup table (interval where the colour occurs) is ignored; i.e. equal spacing intervals are assumed
}
\examples{
\dontrun{
spectral.colors <- lut_to_palette(system.file("luts/spectral", package = "RMINC"))
spectral.colors(100)
}
}
