% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vis3D.R
\name{map_to_colours}
\alias{map_to_colours}
\title{Generate a vector of colours from a map}
\usage{
map_to_colours(colour_map, colour_range = NULL, colour_default = "grey",
  symmetric = NULL, labels = FALSE, palette = heat.colors(255))
}
\arguments{
\item{colour_map}{either a vector with a label/measure/statistic for every vertex
or a character file path pointing to a file with such a vector in a rowwise format.}

\item{colour_range}{a two element numeric vector indicating the min and max values of 
allowable labels/measures/statistics to be includedon the surface}

\item{colour_default}{The colour given to vertices excluded by colour_range}

\item{symmetric}{Whether to have a positive and negative colour scale (not yet implemented)}

\item{labels}{Whether or not the colour_map is a set of discrete labels}

\item{palette}{A palette, AKA look-up-table, providing a linear colour scale for the colours in
\code{colour_map}}
}
\value{
A vector of colours with types corresponding to your input palette
}
\description{
Generate a vector of colours from a map
}
