% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{maskFile}
\alias{maskFile}
\alias{maskFile<-}
\title{Get or set a mask file}
\usage{
maskFile(x, strict = TRUE)

maskFile(x) <- value
}
\arguments{
\item{x}{The minc object}

\item{strict}{Whether or not to throw an error if the mask does not exist}

\item{value}{The replacement value for the mask attribute}
}
\value{
The mask name if getting, or the object invisibly if setting
}
\description{
Mask files control which voxels to perform operations on
}
\section{Functions}{
\itemize{
\item \code{maskFile<-}: setter
}}

