% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{minc.dimensions.sizes}
\alias{minc.dimensions.sizes}
\title{Dimension Sizes}
\usage{
minc.dimensions.sizes(filename)
}
\arguments{
\item{filename}{The filename of the MINC2 volume whose dimension sizes are
to be returned.}
}
\description{
Returns the dimension sizes of a MINC2 volume
}
\seealso{
minc.get.volume
}

