\name{minc.get.volume}
\alias{minc.get.volume}
\title{Read a MINC volume into a 1D array}
\description{Load a 3-dimensional MINC2 volume and returns it as a 1D
  array}
\usage{buffer <- minc.get.volume("filename.mnc")}
\arguments{
  \item{filename}{A string corresponding to the location of the MINC2
    file to be read.}
}
\keyword{file}
