% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{minc.separation.sizes}
\alias{minc.separation.sizes}
\title{Slice Separations}
\usage{
minc.separation.sizes(filename)
}
\arguments{
\item{filename}{A minc file of interest}
}
\value{
a vector of the 3 slice separations
}
\description{
Return the slice separation sizes for determining voxel volume
}
