% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{mincApplyRCPP}
\alias{mincApplyRCPP}
\title{Perform Arbitrary calculations on a collection of mincVolumes}
\usage{
mincApplyRCPP(filenames, fun, ..., mask = NULL, maskval = NULL,
  filter_masked = FALSE, collate = identity)
}
\description{
An RCPP variant of \link{mincApply}, the primary advantage being
that functions of an arbitrary number of arguments can be passed
to mincApplyRCPP.
}

