% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vis2D.R
\name{mincArray}
\alias{mincArray}
\title{A utility function to give a MINC object spatial dimensions}
\usage{
mincArray(volume, dimIndex = 1)
}
\arguments{
\item{volume}{The input volume (from mincLm, mincGetVolume, etc.)}

\item{dimIndex}{The index into a multidimensional object}
}
\value{
A matrix with 3 dimensions
}
\description{
Currently the plotting functions (mincPlotAnatAndStatsSlice) need an object 
with the correct spatial dimensions assigned. While this might in the future happen
at the time those objects are created, for the moment this utility function
works with older style RMINC objects and extract what you need.
}
\note{
R uses Fortran indexing, so dimension assignment is c(dim[3], dim[2], dim[1])
once dimensions are obtained from any libminc functions (which use C indexing)
}
\examples{
\dontrun{
vol <- mincGetVolume("somefile.mnc")
volWithDims <- mincArray(vol)

vs <- mincLm(jacobians ~ genotype, gf)
tvol <- mincArray(vs, 6)
}
}

