% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{mincAttributes}
\alias{mincAttributes}
\alias{setMincAttributes}
\title{Get and Set Minc Specific Attributes}
\usage{
mincAttributes(minc_object)

setMincAttributes(minc_object, updated_attrs)
}
\arguments{
\item{minc_object}{the RMINC object of interest typically a \code{mincMultiDim} or
a related class}

\item{updated_attrs}{A named list containing the new attributes to be replaced}
}
\value{
\code{mincAttributes} returns an attribute list, \code{setMincAttributes} returns 
the updated object
}
\description{
Manage auxillary information contained in RMINC objects. Currently returns all attributes
except dimension information
}
\section{Functions}{
\itemize{
\item \code{setMincAttributes()}: setter

}}
