% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualization.R
\name{mincContour}
\alias{mincContour}
\title{Draw contour lines from a MINC volume}
\usage{
mincContour(volume, dimension = 2, slice = NULL, ...)
}
\arguments{
\item{volume}{the output of mincArray}

\item{dimension}{the dimension (from 1 to 3)}

\item{slice}{the slice number}

\item{...}{other parameters to pass on to \code{\link{contour}}}
}
\description{
Draw contour lines from a MINC volume
}
\examples{
\dontrun{
mincImage(mincArray(anatVol), slice=100, col=gray.colors(255))
mincContour(mincArray(anatVol), slice=100, add=T, col=rainbow(2), levels=c(1000, 1400))
}
}

