% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minc_peaks.R
\name{mincConvertTagToMincArrayCoordinates}
\alias{mincConvertTagToMincArrayCoordinates}
\title{convert tags from world coordinates to mincArray coordinates}
\usage{
mincConvertTagToMincArrayCoordinates(tags, filename)
}
\arguments{
\item{tags}{from mincGetTagFile}

\item{filename}{path to a MINC file to obtain coordinate info from}
}
\value{
tags as a ntags by 4 matrix
}
\description{
convert tags from world coordinates to mincArray coordinates
}

