% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{mincConvertVoxelMatrix}
\alias{mincConvertVoxelMatrix}
\title{Convert Voxel to World Coordinates}
\usage{
mincConvertVoxelMatrix(filename, voxel_matrix)
}
\arguments{
\item{filename}{The minc file indicating a coordinate grid}

\item{voxel_matrix}{a 3xN matrix of voxel coordinates}
}
\value{
a 3xN matrix of world coordinates
}
\description{
Convert a 3xN matrix of voxel coordinates to world coordinates
with respect to a given minc file.
}

