% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{mincConvertVoxelToWorld}
\alias{mincConvertVoxelToWorld}
\title{Voxel to World}
\usage{
mincConvertVoxelToWorld(filename, v1, v2, v3)
}
\arguments{
\item{filename}{Character vector with paths to minc files}

\item{v1}{First voxel coordinate}

\item{v2}{Second voxel coordinate}

\item{v3}{Third voxel coordinate}
}
\value{
a 3-component numeric vector of world coordinates
}
\description{
Convert a discrete set of voxel coordinates into the volumes
continous world coordinate space
}
