% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{mincConvertWorldMatrix}
\alias{mincConvertWorldMatrix}
\title{Convert World to Voxel Coordinates}
\usage{
mincConvertWorldMatrix(filename, world_matrix, nearest_voxel = TRUE)
}
\arguments{
\item{filename}{The minc file indicating a coordinate grid}

\item{world_matrix}{a 3xN matrix of world coordinates}

\item{nearest_voxel}{logical whether to round the results to the nearest voxel}
}
\value{
a 3xN matrix of voxel coordinates
}
\description{
Convert a 3xN matrix of world coordinates to voxel coordinates
with respect to a given minc file.
}
