% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mincCorrelation}
\alias{mincCorrelation}
\title{Perform a correlation between a set of minc volumes and another variable.}
\usage{
mincCorrelation(filenames, grouping, mask = NULL, maskval = NULL)
}
\arguments{
\item{filenames}{Filenames of the MINC volumes across which to run the correlation}

\item{grouping}{Contains same number of elements as
filenames; contains values with which to correlate}

\item{mask}{A mask specifying which voxels are to be included in the correlation}

\item{maskval}{The value with which to mask the data (data will masked +/- 0.5 around this value}
}
\value{
The output will be a single vector containing as many
elements as there are voxels in the input files, with that voxel's correlation value (Pearson
correlation coefficient)
}
\description{
Perform a correlation between a set of minc volumes.
}
\examples{
getRMINCTestData()
gf <- read.csv("/tmp/rminctestdata/minc_summary_test_data.csv")
mc <- mincCorrelation(gf$jacobians_0.2,gf$Weight);
}

