% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mincFDR.mincLmer}
\alias{mincFDR.mincLmer}
\title{mincFDR for the output of mincLmer}
\usage{
\method{mincFDR}{mincLmer}(buffer, mask = NULL)
}
\description{
same as mincFDR, except adds warnings based on difficulty in estimating
degrees of freedom for mixed effects models
}

