% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mincFDR.mincLogLikRatio}
\alias{mincFDR.mincLogLikRatio}
\title{mincFDR for the output of mincLogLikRatio}
\usage{
\method{mincFDR}{mincLogLikRatio}(buffer, mask = NULL)
}
\description{
Very similar to the default mincFDR, except that it adds
the ability to correct the p-values based on having run
the parametric bootstrap first.
}
\examples{
\dontrun{
test <- mincLmer(jacobians ~ day + (1|mouse), gf, mask="/micehome/jlerch/R-tests/small-mask.mnc",REML=F)
test2 <- mincLmer(jacobians ~ ns(day,2) + (1|mouse), gf, mask="/micehome/jlerch/R-tests/small-mask.mnc",REML=F)
out <- mincLogLikRatio(test, test2)
# compute the FDR just using the chi squared approximation
mincFDR(out)
out <- mincLogLikRatioParametricBootstrap(out, nvoxels=25, nsims=500)
# also compute the FDR with corrected p-values based on the chi-squared approximation
mincFDR(out)
}
}

