% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mincFDRMask}
\alias{mincFDRMask}
\title{returns the specified mask, the mask associated with the buffer, or a vector of ones}
\usage{
mincFDRMask(mask, buffer)
}
\description{
returns the specified mask, the mask associated with the buffer, or a vector of ones
}

