% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mincFDRThresholdVector}
\alias{mincFDRThresholdVector}
\title{a utility function to compute thresholds}
\usage{
mincFDRThresholdVector(pvals, qvals, thresholdFunc = NULL,
  p.thresholds = c(0.01, 0.05, 0.1, 0.15, 0.2))
}
\arguments{
\item{pvals}{a vector of pvalues}

\item{qvals}{a vector of corrected qvalues (such as returend by p.adjust)}

\item{thresholdFunc}{a function that returns the threshold given a vector of pvalues}

\item{p.thresholds}{the pvalues at which to compute the threshold

The function should be the quantile function for the distribution being tested. For example,
for the chi squared distribution the function would be:
tfunc <- function(x) { qchisq(max(x), df[[i]], lower.tail=FALSE) }}
}
\description{
a utility function to compute thresholds
}

