% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{mincFileCheck}
\alias{mincFileCheck}
\title{Minc File Check}
\usage{
mincFileCheck(filenames)
}
\arguments{
\item{filenames}{paths to the files to check}
}
\value{
Returns NULL invisibly
}
\description{
test to see whether files exist and are readable
}
\details{
Throws an error if a file is not found or not readable
}
