% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minc_peaks.R
\name{mincFindPeaks}
\alias{mincFindPeaks}
\title{Finds peak values in a MINC file or volume}
\usage{
mincFindPeaks(inputStats, column = 1, direction = "both",
  minDistance = NA, threshold = NA, posThreshold = NA,
  negThreshold = NA)
}
\arguments{
\item{inputStats}{Either an RMINC object or a filename from which peaks are to be determined.}

\item{column}{If an RMINC object then the column name or number}

\item{direction}{"both", "pos" for positive peaks only, or "neg" for negative peaks only}

\item{minDistance}{minimum distance, in mm, between peaks}

\item{threshold}{threshold above which to consider peaks}

\item{posThreshold}{positive threshold (overrides threshold if both are given)}

\item{negThreshold}{negative threshold (overrides threshold if both are given)}
}
\value{
The set of peaks as a matrix
}
\description{
Inputs a MINC file or RMINC output and identifies peak voxels (given some constraints) within
}
\details{
Provides an interface to find_peaks, a command line program that comes with the MINC library.
}

