% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{mincGetMask}
\alias{mincGetMask}
\title{Minc Masks}
\usage{
mincGetMask(mask)
}
\arguments{
\item{mask}{Either the path to a mask file or a numeric vector representing the mask}
}
\value{
a numeric mask vector
}
\description{
Either create a mask from a file name (by reading the volume)
or pass the vector along undisturbed.
}
