% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_peaks.R
\name{mincGetTagFile}
\alias{mincGetTagFile}
\title{read a tag file}
\usage{
mincGetTagFile(filename)
}
\arguments{
\item{filename}{string containing filename to read}
}
\value{
tags as a ntags by 7 matrix
}
\description{
read a tag file
}
