% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{mincGetVector}
\alias{mincGetVector}
\title{Voxel Vector}
\usage{
mincGetVector(filenames, v1, v2, v3, v.length = NULL)
}
\arguments{
\item{filenames}{Character vector with paths to minc files}

\item{v1}{First voxel coordinate}

\item{v2}{Second voxel coordinate}

\item{v3}{Third voxel coordinate}

\item{v.length}{The number of values to return}
}
\description{
Given a voxel coordinate, return the value at those coordinates
from a set of minc files as a numeric vector
}
