% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{mincGetVoxel}
\alias{mincGetVoxel}
\title{Retrieve Voxel Values}
\usage{
mincGetVoxel(filenames, v1, v2 = NULL, v3 = NULL)
}
\arguments{
\item{filenames}{paths to the minc files}

\item{v1}{Either a 3-element vector of voxel coordinates or the first}

\item{v2}{the second voxel coordinate if not NULL}

\item{v3}{the third voxel coordinate if not NULL}
}
\value{
Returns a \code{mincVoxel} object containing a vector
of intensities and attributes specify the voxel and world coordinates
of the values.
}
\description{
Return the intensity of a given voxel in a set of minc files
}

