% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{mincGetWorldVoxel}
\alias{mincGetWorldVoxel}
\title{World Vector}
\usage{
mincGetWorldVoxel(filenames, v1, v2 = NULL, v3 = NULL)
}
\arguments{
\item{filenames}{A character vector of one or more filenames}

\item{v1}{First world coordinate}

\item{v2}{Second world coordinate}

\item{v3}{Third world coordinate}
}
\value{
a vector of values
}
\description{
Given a set of world coordinates return the value at those coordinates
from a set of minc files as a numeric vector
}

