\name{mincIO.printMincInfo}
\Rdversion{1.1}
\alias{mincIO.printMincInfo}
\title{    Print a MincInfo Object}

\description{This function formats and prints a MincInfo type object.}

\usage{
mincIO.printMincInfo(mincInfo)
}

\arguments{
  \item{mincInfo}{An object of type MincInfo.}
}

\details{This routine is generally only called by the \code{print} or \code{show} generic functions
         whenever a MincInfo object needs to be printed.}

\value{This function is run for its side-effect.}

\author{Jim Nikelski \email{nikelski@bic.mni.mcgill.ca}}

\seealso{
    \code{\link{mincIO.print-methods}} \code{\linkS4class{MincInfo}} 
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{internal}

