% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_peaks.R
\name{mincLabelPeaks}
\alias{mincLabelPeaks}
\title{label peaks with the name of the atlas structure they are in}
\usage{
mincLabelPeaks(peaks, atlas, defs = getOption("RMINC_LABEL_DEFINITIONS"))
}
\arguments{
\item{peaks}{the output of \code{\link{mincFindPeaks}}}

\item{atlas}{the atlas volume, either as a \code{\link{mincArray}} or as a filename}

\item{defs}{the atlas definitions, same as used for \code{\link{anatGetAll}}}
}
\value{
the peaks data frame, with an extra column containing the label
}
\description{
label peaks with the name of the atlas structure they are in
}
\examples{
\dontrun{
peaks <- mincFindPeaks(-log10(qvs), "Neonatal:time.to.sac", 
                       "pos", posThreshold=1.3, minDistance=1)
peaks <- mincLabelPeaks(peaks, 
                         atlasVol, 
                         defs="Dorr_2008_Steadman_2013_Ullmann_2013_mapping_of_labels.csv")
}
}

