% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mincLm}
\alias{mincLm}
\title{Linear model at Every Voxel}
\usage{
mincLm(formula, data = NULL, subset = NULL, mask = NULL, maskval = NULL)
}
\arguments{
\item{formula}{The linear model formula. The left-hand term consists of the MINC filenames over which to compute the models at every voxel.The RHS of the formula may contain one term with filenames. If so only the + operator may be used, and only two terms may appear on the RHS}

\item{data}{The dataframe which contains the model terms.}

\item{subset}{Subset definition.}

\item{mask}{Either a filename or a vector of values of the same length as the input files. The linear model will only be computed
inside the mask.}
}
\value{
mincLm returns a mincMultiDim object which contains a series of columns corresponding to the terms in the linear model. The first
column is the F-statistic of the significance of the entire volume, the following columns contain the marginal t-statistics for each of the terms in
the model
}
\description{
Linear Model at Every Voxel
}
\details{
This function computes a linear model at every voxel of a set of files. The function is a close cousin to lm, the key difference
being that the left-hand side of the formula specification takes a series of filenames for MINC files.
}
\examples{
getRMINCTestData()
# read the text file describing the dataset
gf <- read.csv("/tmp/rminctestdata/test_data_set.csv")
# Compute a linear model at each voxel
vs <- mincLm(jacobians_fixed_2 ~ Sex, gf)
}
\seealso{
mincWriteVolume,mincFDR,mincMean, mincSd
}

