% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mincLmer}
\alias{mincLmer}
\title{mincified version of lmer from lme4}
\usage{
mincLmer(formula, data, mask = NULL, parallel = NULL, REML = TRUE,
  control = lmerControl(), start = NULL, verbose = 0L)
}
\arguments{
\item{formula}{the lmer formula, filenames go on left hand side}

\item{data}{the data frame, all items in formula should be in here}

\item{mask}{the mask within which lmer is solved}

\item{parallel}{how many processors to run on (default=single processor).
Specified as a two element vector, with the first element corresponding to
the type of parallelization (sge or snowfall), and the second to the number
of processors to use.}

\item{REML}{whether to use use Restricted Maximum Likelihood or Maximum Likelihood}

\item{control}{lmer control function}

\item{start}{lmer start function}

\item{verbose}{lmer verbosity control}
}
\value{
a matrix where rows correspond to number of voxels in the file and columns to number of terms in the formula
}
\description{
mincLmer should be used the same way as a straight lmer call, except
that the left hand side of the equation contains minc filenames rather than
an actual response variable.
}
\examples{
\dontrun{
vs <- mincLmer(filenames ~ age + sex + (age|id), data=gf, mask="mask.mnc")
mincWriteVolume(vs, "age-term.mnc", "tvalue-age")
# run in parallel with multiple processors on the local machine
vs <- mincLmer(filenames ~ age + sex + (age|id), data=gf, mask="mask.mnc", parallel=c("snowfall", 4))
# run in parallel with multiple processors over the sge batch queueing system
vs <- mincLmer(filenames ~ age + sex + (age|id), data=gf, mask="mask.mnc", parallel=c("sge", 4))
}
}
\seealso{
\code{\link{lmer}} for description of lmer and lmer formulas; \code{\link{mincLm}}
}

