% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_voxel_statistics.R
\name{mincPairedTtest}
\alias{mincPairedTtest}
\title{Minc Paired T Test}
\usage{
mincPairedTtest(filenames, grouping, mask = NULL, maskval = NULL)
}
\arguments{
\item{filenames}{Filenames of the MINC volumes across which to run the t-test}

\item{grouping}{Contains same number of elements as
filenames; must contain exactly two groups with which to compare means. The two groups
must be the same length.}

\item{mask}{A mask specifying which voxels are to be included in the test}

\item{maskval}{The value with which to mask the data (data will masked +/- 0.5 around this value}
}
\value{
The output will be a single vector containing as many
         elements as there are voxels in the input files, with that voxel's t-statistic
}
\description{
Perform a paired t-test across a set of minc volumes
}
\examples{
\dontrun{ 
getRMINCTestData() 
gf <- read.csv("/tmp/rminctestdata/minc_summary_test_data.csv") 
gf = gf[1:20,]
mptt <- mincPairedTtest(gf$jacobians_0.2,gf$Strain)
}
}
