% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vis2D.R
\name{mincPlotAnatAndStatsSlice}
\alias{mincPlotAnatAndStatsSlice}
\title{Anatomy and Statistics Slice}
\usage{
mincPlotAnatAndStatsSlice(anatomy, statistics, slice = NULL, dimension = 2,
  low = min(statistics, na.rm = TRUE), high = max(statistics, na.rm = TRUE),
  anatLow = min(anatomy, na.rm = TRUE), anatHigh = max(anatomy, na.rm =
  TRUE), symmetric = FALSE, col = NULL, rcol = NULL, legend = NULL)
}
\arguments{
\item{anatomy}{A minc array of the anatomy volume to plot}

\item{statistics}{optional statistics or label file to overlay on anatomy slices}

\item{slice}{the voxel index of the slice of interest}

\item{dimension}{integer denoting which dimension to slice across}

\item{low}{the minimum statistic to plot}

\item{high}{the maximum statistic to plot}

\item{anatLow}{the minimum anatomy intensity to plot}

\item{anatHigh}{the maximum antomy intensity to plot}

\item{symmetric}{whether the statistics are symmetric (such as for t-statistics)}

\item{col}{colours for statistics}

\item{rcol}{colours for negative statistics if using a symmetric statistic}

\item{legend}{an optional string to name the legend, indicating desire for a legend
(or not)}
}
\value{
invisible NULL
}
\description{
Plot a given slice through a MINC volume and superimpose statistics
on the slice.
}

