% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vis2D.R
\name{mincPlotPeak}
\alias{mincPlotPeak}
\title{Plotting of peaks}
\usage{
mincPlotPeak(peak, anatomy, statistics, dim = 2, crossCol = "green",
  crossSize = 4, plotFunction = NULL, ...)
}
\arguments{
\item{peak}{a row from \code{\link{mincFindPeaks}}}

\item{anatomy}{a mincArray for the underlying anatomy}

\item{statistics}{a mincArray for the stats volume}

\item{dim}{the dimension (1:3)}

\item{crossCol}{the colour for the cross-hair}

\item{crossSize}{the size (cex) of the cross-hair}

\item{plotFunction}{a function which will produce a graph}

\item{...}{other details to pass on to \code{\link{mincPlotAnatAndStatsSlice}}}
}
\description{
Plots a slice containing a peak. Optionally plots a graph of
that peak alongside.
}
\examples{
\dontrun{
peaks <- mincFindPeaks(-log10(qvs), "Neonatal:time.to.sac", "pos",
posThreshold=1.3, minDistance=1)
p <- function(peak) {
  gfTiming$voxel <- mincGetWorldVoxel(gfTiming$reljacobians02,
                                      peak["x"], peak["y"], peak["z"])
  qplot(time.to.sac, exp(voxel), data=gfTiming, colour=Neonatal,
        geom="boxplot") + theme_classic()
}
mincPlotPeak(peaks[1,], anatVol, -log10(mincArray(qvs, "Neonatal:time.to.sac")), 
             anatLow=700, anatHigh=1400, low=1, high=4, col=heat.colors(244), 
             crossCol = "blue", crossSize = 3, plotFunction = p)
}
}

