% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mincPlotSliceSeries}
\alias{mincPlotSliceSeries}
\title{Minc Slice Series}
\usage{
mincPlotSliceSeries(anatomy, statistics = NULL, dimension = 2,
  mfrow = c(4, 5), low = NULL, high = NULL, anatLow = NULL,
  anatHigh = NULL, col = ifelse(symmetric, heat.colors(255),
  colorRampPalette(c("red", "yellow"))(255)), begin = 1,
  end = (dim(anatomy)[dimension] - 1), symmetric = FALSE, legend = NULL,
  locator = !is.null(legend), plottitle = NULL, indicatorLevels = NULL,
  discreteStats = FALSE, flip = FALSE)
}
\arguments{
\item{anatomy}{A minc array of the anatomy volume to plot}

\item{statistics}{optional statistics or label file to overlay on anatomy slices}

\item{dimension}{integer denoting which dimension to slice across}

\item{low}{the minimum intensity to plot, calculated from the histogram if NULL}

\item{high}{the maximum intensity to plot, calculated from the histogram if NULL}

\item{underTransparent}{boolean whether or not the background should be transparent
or the lowest col}

\item{begin}{the first slice to plot, defaults to 1}

\item{end}{the last slice to plot, defaults to the last slice}

\item{plottitle}{the title of the plot if desired}

\item{legend}{an optional string to name the legend, indicating desire for a legend
(or not)}

\item{boolean}{whether or not to draw the locator, defaults to whether or not
you requested a legend}

\item{indicatorLevels}{numeric vector indicating where to draw slice lines on the
locator, defaults to every slice}

\item{flip}{boolean, whether to invert the y-axis.
On certain systems \link{mincImage} produces vertically reflected images, use this switch
to fix that.}
}
\description{
Plot a simple series of slices through a minc volume
on a given dimension. Optionally include a locator
contour to show where slices are.
}

