% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mincSummaries}
\alias{mincMean}
\alias{mincSd}
\alias{mincSum}
\alias{mincSummaries}
\alias{mincVar}
\title{Create descriptive statistics across a series of MINC volumes}
\usage{
mincMean(filenames, grouping = NULL, mask = NULL, maskval = NULL)
}
\arguments{
\item{filenames}{Filenames of the MINC volumes across which to create the
descriptive statistic.}

\item{grouping}{Optional grouping - contains same number of elements as
filenames; the results will then have the descriptive
statistic computed separately for each group.}

\item{mask}{A mask specifying which voxels are to be included in the
summary.}
}
\value{
The output will be a single vector containing as many
         elements as there are voxels in the input files. If a
         grouping factor was specified then the output will be a
         matrix consisiting of as many rows as there were voxels in
         the files, and as many columns as there were groups.
}
\description{
This function is used to compute the mean, standard deviation,
   		sum, or variance of every region in a set of MINC volumes.
		An optional grouping variable will split the computation by group
    		rather than performing it across all volumes as is the default.
}
\examples{
getRMINCTestData()
gf <- read.csv("/tmp/rminctestdata/minc_summary_test_data.csv")
mm <- mincMean(gf$jacobians_0.2);
ms <- mincSd(gf$jacobians_0.2);
mv <- mincVar(gf$jacobians_0.2,gf$Strain);
ms2 <- mincSum(gf$jacobians_0.2,gf$Strain);
}

