% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{mincTable}
\alias{mincTable}
\title{Read a collection of minc volumes}
\usage{
mincTable(filenames, mask = NULL, file_backed = FALSE, ...)
}
\arguments{
\item{filenames}{A character vector of minc filenames. All minc files must have the
same shape.}

\item{mask}{Either a character vector with a path to a mask file, a mincVolume, or a
numeric/logical vector indicating which voxels to include in the table.}

\item{file_backed}{logical, whether to use a file backed matrix for storing the table.}

\item{...}{Additional arguments to `bigstatr::FBM`}
}
\value{
If `file_backed` is `FALSE` return a matrix with ncol equal to the number of files.
The number of rows is either the product of the volume dimensions, or if a mask is supplied,
the number of voxels where the mask > 0.5. If `file_backed` is `TRUE` return an `FBM` object
from the `bigstatr` package.
}
\description{
This reads a collection of `mincVolumes` into an optionally file-backed matrix
}
