% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_voxel_statistics.R
\name{mincTtest}
\alias{mincTtest}
\title{Minc T-test}
\usage{
mincTtest(filenames, grouping, mask = NULL, maskval = NULL)
}
\arguments{
\item{filenames}{Filenames of the MINC volumes across which to run the t-test}

\item{grouping}{Contains same number of elements as
filenames; must contain exactly two groups with which to compare means}

\item{mask}{A mask specifying which voxels are to be included in the test}

\item{maskval}{The value with which to mask the data (data will masked +/- 0.5 around this value}
}
\value{
The output will be a single vector containing as many
         elements as there are voxels in the input files, with that voxel's t-statistic
}
\description{
Perform an unpaired,unequal variance t-test across a set of minc volumes
}
\examples{
\dontrun{ 
getRMINCTestData() 
gf <- read.csv("/tmp/rminctestdata/minc_summary_test_data.csv") 
mtt <- mincTtest(gf$jacobians_0.2,gf$Strain)
} 
}

