% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mincVectorToVoxelCoordinates}
\alias{mincVectorToVoxelCoordinates}
\title{converts a vector index to the voxel indices in MINC}
\usage{
mincVectorToVoxelCoordinates(volumeFileName, vectorCoord)
}
\arguments{
\item{volumeFileName}{the filename of the MINC volume}

\item{vectorCoord}{the integer array index to convert}
}
\value{
a vector of length 3 containing the MINC indices in volume dimension order
}
\description{
RMINC stores volume data as 1d arrays. This function gives the
corresponding voxel coordinates (in the dimension order of the volume)
for an index into the 1d array.
}
\examples{
\dontrun{
index <- mincVectorToVoxelCoordinates("filename.mnc", 345322)
voxel <- mincGetVoxel(gf$filenames, index)
}
}

