% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mincWilcoxon}
\alias{mincWilcoxon}
\title{Perform a Mann-Whitney U  between a set of minc volumes.}
\usage{
mincWilcoxon(filenames, grouping, mask = NULL, maskval = NULL)
}
\arguments{
\item{filenames}{Filenames of the MINC volumes across which to run the test}

\item{grouping}{Contains same number of elements as
filenames; must contain exactly two groups.}

\item{mask}{A mask specifying which voxels are to be included in the test}

\item{maskval}{The value with which to mask the data (data will masked +/- 0.5 around this value}
}
\value{
The output will be a single vector containing as many
elements as there are voxels in the input files, with that voxel's U value (lower one).
The number of observations in each sample is also saved as an attribute(m and n) so the result
can be passed into mincFDR.
}
\description{
Perform a Mann-Whitney U test between a set of minc volumes.
}
\examples{
getRMINCTestData()
gf <- read.csv("/tmp/rminctestdata/minc_summary_test_data.csv")
mw <- mincWilcoxon(gf$jacobians_0.2,gf $Strain);
}

