% Generated by roxygen2 (4.0.1): do not edit by hand
\name{obj_montage}
\alias{obj_montage}
\title{Brain Montage}
\usage{
obj_montage(left_obj, right_obj, left_map, right_map, output = NULL, ...,
  add_normals = TRUE, colour_title = "", close_on_output = TRUE)
}
\arguments{
\item{left_obj}{A \code{bic_obj} probably created by \link{read_obj} with the left hemisphere
of a subject's brain}

\item{right_obj}{A \code{bic_obj} probably created by \link{read_obj} with the right hemisphere
of a subject's brain}

\item{left_map}{a colour map to apply to the left hemisphere see \link{colour_mesh} for details}

\item{right_map}{a colour map to apply to the right hemisphere see \link{colour_mesh} for details}

\item{output}{Either NULL or a file path to write the snapshot.}

\item{...}{additonal parameters to be passed to \link{create_mesh}}

\item{add_normals}{Whether or not to add normals to the surface objects, see \link{create_mesh} for
details}

\item{close_on_output}{Whether or not to close the output after taking a snapshot, defaults to
TRUE}
}
\value{
invisible NULL
}
\description{
Plot the left and right hemispheres of brain colourized by two colour maps
with a colour bar in the middle
}
\details{
This function is designed to do a simple 6 angle plot for statistic maps of the left and right
hemispheres of subject's brain. It defaults to leaving the rgl device open so that you can
take a snapshot after tweaking the angles, or changing the colour bar with \link{add_colour_bar}.
Its other mode is to take a snapshot after it has finished adding the 3d objects to the scene,
in this mode, whether or not to keep the window open can be configured with close_on_output.
}

