% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vis3D.R
\name{obj_to_graph}
\alias{obj_to_graph}
\title{Object to Graph}
\usage{
obj_to_graph(bic_obj)
}
\arguments{
\item{bic_obj}{The \code{bic_obj} object, generated by \link{read_obj} to be converted}
}
\value{
an \code{igraph} object representing the object's mesh
}
\description{
Convert a bic_obj object to a graph for handy graph algorithms
}
