% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vis3D.R
\name{plot.bic_lines}
\alias{plot.bic_lines}
\title{Plot A bic_lines object}
\usage{
\method{plot}{bic_lines}(x, dimension = 2, ...)
}
\arguments{
\item{x}{an \code{bic_lines} object}

\item{dimension}{which axis to display the lines on}

\item{...}{additional parameters to pass to \link{segments}}
}
\value{
NULL invisibly
}
\description{
Add lines corresponding to the coordinates in a bic_lines
object to a figure. Only draws the projection of the lines
on a single dimension, no regard is given for whether the
lines are near the slice of interest.
}
