% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_functions.R
\name{plot.obj_mesh}
\alias{plot.obj_mesh}
\title{Plot an BIC obj mesh}
\usage{
plot.obj_mesh(x, colour_bar = TRUE, ...)
}
\arguments{
\item{x}{a \link{obj_mesh} object}

\item{colour_bar}{whether or not to add a colour bar}

\item{additional}{parameters to pass to add_colour_bar}
}
\value{
returns x invisibly
}
\description{
Create a plot of BIC obj_mesh, potentially colourized by \link{colour_mesh}
and potentially including a colour bar
}

