% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_bic_obj_overlay}
\alias{plot_bic_obj_overlay}
\title{Plot a BIC obj with a colour map}
\usage{
plot_bic_obj_overlay(bic_obj, colour_map, palette = heat.colors(255),
  colour_bar = TRUE, ...)
}
\arguments{
\item{bic_obj}{A \code{bic_obj} probably created by \link{read_obj}}

\item{colour_map}{A numeric vector equal in length to the number of vertices
in the \code{bic_obj} or the path to a text file with one line per vertex with
colour information.}

\item{A}{palette, AKA look-up-table, providing a linear colour scale for the colours in
\code{colour_map}}

\item{...}{additional arguments to \link{create_mesh} including but not limited
to colour, specular, and add_normals}

\item{invisibly}{return the colourized mesh object}
}
\description{
Create a colour mapped bic_obj and plot it in the current rgl
device, opening a new one if necessary.
}

