% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{pt2}
\alias{pt2}
\title{Two tailed version of pt}
\usage{
pt2(q, df, log.p = FALSE)
}
\arguments{
\item{q}{a vector of quantiles}

\item{df}{degrees of freedom}

\item{log.p}{whether to return the natural logarithm of p}
}
\value{
A vector of probabilities
}
\description{
Generate two tailed probabilities from
the cummulative t distribution function with
a given number of degrees of freedom.
}
\seealso{
pt
}
