% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_parallel.R
\name{qMincApply}
\alias{qMincApply}
\alias{qMincMap}
\alias{qMincReduce}
\alias{qMincRegistry}
\title{True cluster mincApply}
\usage{
qMincApply(filenames, fun, ..., mask = NULL, batches = 4,
  tinyMask = FALSE, slab_sizes = NULL, parallel_method = c("none", "sge",
  "pbs", "multicore", "interactive", "custom"), cluster_functions = NULL,
  resources = list(), packages = c("RMINC"), temp_dir = tempdir(),
  registry_name = "qMincApply_registry", registry_dir = file.path(temp_dir,
  registry_name), cores = NULL, wait = TRUE, cleanup = TRUE,
  clobber = FALSE, collate = simplify2minc)

qMincRegistry(registry_name = "qMincApply_registry",
  parallel_method = c("none", "sge", "pbs", "multicore", "interactive",
  "custom"), packages = c("RMINC"), cluster_functions = NULL,
  registry_dir = file.path(tempdir(), registry_name), cores = NULL,
  clobber = FALSE)

qMincMap(registry, filenames, fun, ..., mask = NULL, slab_sizes = NULL,
  parallel_method = c("none", "sge", "pbs", "multicore", "interactive",
  "custom"), batches = 4, tinyMask = FALSE, temp_dir = tempdir(),
  resources = list(), cores = NULL)

qMincReduce(registry, ignore_incompletes = FALSE, wait = FALSE,
  collate = simplify2minc)
}
\arguments{
\item{filenames}{Paths to the minc files to apply accross}

\item{fun}{An arbitrary R function to be applied}

\item{...}{extra arguments to pass down through \code{qMincMap} to \code{mcMincApply}
to \code{mincApplyRCPP} and finally to \code{fun}, there is a chance arguments 
here will be trapped by one of the functions on this chain, when in doubt partially
apply \code{fun} to its arguments before hand and do not use positional arguments, they
are almost certainly not going to work as expected.}

\item{mask}{The mask used to select voxels to apply to}

\item{batches}{The number of batches to divide the job into, this is ignored for
multicore jobs, with the number of batches set to the number of cores.}

\item{tinyMask}{Shrink the mask for testing}

\item{slab_sizes}{A 3 element vector indicating large a chunk of data to read from each minc file 
at a time defaults to one slice along the first dimension.}

\item{parallel_method}{Where to run the batches, defaults to \code{"none"} which results
in using the currently loaded configuration for BatchJobs. The other options
are \code{"sge"} for Sun Grid Engine, \code{"pbs"} for Torque, \code{"multicore"} for
local (not recommended, use \link{mcMincApply}), \code{"interactive"} for sequential local
processing, and \code{"custom"} in which case a cluster.functions object must be
passed in through \code{cluster_functions} (also not recommend, for more control 
use the .BatchJobs.R configuration system, see details).}

\item{cluster_functions}{Custom cluster functions to use if parallel_method = "custom"}

\item{resources}{The resources to request for each job, overrides the default.resources
specified in .BatchJobs.R config files (see details). These include things like vmem, 
walltime, and nodes.}

\item{packages}{packages to be loaded for each job in a registry}

\item{temp_dir}{A directory to store files needed for the parallelization
and job management}

\item{registry_name}{The name to give your BatchJobs registry}

\item{registry_dir}{The path for your BatchJobs registry, defaults to
\code{temp_dir/registry_name}}

\item{cores}{the number of cores to use, in all cases except multicore this
defaults to 1, with a reminder message if method "none" or "custom" are used,
for multicore it defaults to \code{max(getOption("mc.cores"), parallel::detectCores() - 1)}}

\item{wait}{Whether to wait for your results or return a registry object
to be checked on later}

\item{cleanup}{Whether to empty the registry after a successful run defaults
to true}

\item{clobber}{Whether to overwrite an existing registry at \code{registry_dir}}

\item{collate}{A function to collate the returned list into another object type.}

\item{registry}{A pre-made BatchJobs registry}

\item{ignore_incompletes}{Whether to reduce the results with \code{qMincReduce}
even if all jobs are not complete.}
}
\value{
\itemize{
\item{If \code{qMincApply} is called with \code{wait = TRUE} or if \code{qMincReduce}
is called, the results are returned after collation with \code{collate}
}
\item{If \code{qMincApply} is called with \code{wait = FALSE} or if \code{qMincRegistry} or
\code{qMincMap} are called  a BatchJobs registry is returned that can be used to 
query job states, kill jobs, and collected results
}
}
}
\description{
Split a minc apply job into batches and process it either locally
or a true grid computing setup. Endeavours to provide an abstract
and customizable interface for job scheduling based on the BatchJobs
package. Basic steps of the apply is to 
\itemize{
\item{create a registry with \link{qMincRegistry} where jobs are coordinated and 
results are deposited}
\item{map a function over batches of voxels in a collection of minc volumes with
\link{qMincMap}, generating submission scripts for the queue scheduler and 
submitting the jobs}
\item{Collect the results from each batch with \code{qMincReduce}, reorganizing the voxel 
results as necessary to reproduce the original order, and then collating the results into a
usable object}
}
Interfaces are provided to perform all three steps at once, either through 
\link{qMincApply} or the more general \link{pMincApply}. By default
\link{qMincApply} will wait for the cluster to finish processing all jobs
although the jobs can be submitted and the R session closed while still
maintaining the ability to access results when the jobs have finished.
}
\details{
RMINC's batching facilities are inherited with little modification from
the BatchJobs package, mostly just providing handy wrappers to handle registry
creation, batching, submission, and reduction. The abstractions provided are very leaky
and it is worth learning about BatchJobs to handle more complex situations. This being
said a high degree of flexibility is already available. RMINC honours the hierarchy
of configuration files <BatchJobs>/.BatchJobs.R < ~/.BatchJobs.R < getwd()/.BatchJobs.R
these files are the place set sensible defaults for your usage. The template scripts
provided for SGE and PBS are located in \code{system.file("parallel/", package = "RMINC")}
as well as some example .BatchJobs.R files.
}
\section{Functions}{
\itemize{
\item \code{qMincRegistry}: registry

\item \code{qMincMap}: map

\item \code{qMincReduce}: reduce
}}
\seealso{
\url{https://www.jstatsoft.org/article/view/v064i11} \link{pMincApply} \link{mcMincApply}
}

