% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vis3D.R
\name{read_line_obj}
\alias{read_line_obj}
\title{Read a BIC-obj line file}
\usage{
read_line_obj(line_obj)
}
\arguments{
\item{line_obj}{Path to the object file of interest}
}
\value{
\code{bic_lines} object, which is a list of matrices, each 
matrix coresponds to one line in the object. The matrices are 3xN matrices 
of world coordinates.
}
\description{
Parse the BIC obj format for when the object contains
lines instead of a mesh.
}

