\name{RMINC-internal}
\title{Internal RMINC objects}
\alias{.First.lib}
\alias{.Last.lib}
\alias{.requireCachedGenerics}

\alias{[.pdMat}
\alias{[.reStruct}
\alias{[<-.pdMat}

% anatomical functions
\alias{anatApply}
\alias{anatFDR}
\alias{anatGetFile}
\alias{anatLm}
\alias{anatRenameRows}

\alias{as.matrix.pdMat}
\alias{as.matrix.reStruct}
\alias{civet.AllROIs}
\alias{civetAllROIs}
\alias{civetFilenames}
\alias{coef.modelStruct}
\alias{corMatrix.pdMat}
\alias{corMatrix.reStruct}
\alias{f}
\alias{fast.qvalue}
\alias{fitted.lme}
\alias{fitted.rmincLmeStruct}
\alias{fixef.lme}
\alias{formula.lme}
\alias{formula.modelStruct}
\alias{formula.pdMat}
\alias{logDet.newpdNatural}
\alias{logDet.pdMat}
\alias{logDet.pdSymm}
\alias{logLik.reStruct}
\alias{logLik.rmincLme}
\alias{logLik.rmincLmeStruct}
\alias{logLik.rmincLmeStructInt}

\alias{minc.get.hyperslab2}
\alias{minc.get.slices}
\alias{minc.get.volumes}
\alias{minc.slice.loop}
\alias{minc.t.test}
\alias{minc.wilcoxon.test}
\alias{mincAnova}
\alias{mincApplyLme}
\alias{mincConvertVoxelToWorld}
\alias{mincConvertWorldToVoxel}
\alias{mincFileCheck}
\alias{mincGetMask}
\alias{mincGetVector}
\alias{mincGetVoxel}
\alias{mincGetWorldVoxel}
\alias{mincSlowLme}
\alias{mincSummary}
\alias{mincWriteVolume.default}

% mincIO internals
\alias{mincIO.makeNewVolumeX}
\alias{mincIO.readByVoxelX}
\alias{mincIO.readVolumeX}
\alias{mincIO.writeVolumeX}

\alias{model.matrix.reStruct}
\alias{needUpdate.modelStruct}
\alias{needUpdate.reStruct}
\alias{newACF}
\alias{newallVarsRec}
\alias{newasOneFormula}
\alias{newasTable}
\alias{newaugPred}
\alias{newBIC}
\alias{newBIC.lm}
\alias{newBIC.logLik}
\alias{newcoef.newpdNatural}
\alias{newcoef.pdMat}
\alias{newcoef.pdSymm}
\alias{newcoef.reStruct}
\alias{newcoef<-}
\alias{newcoef<-.modelStruct}
\alias{newcoef<-.pdMat}
\alias{newcoef<-.reStruct}
\alias{newcoefficients<-}
\alias{newcollapse}
\alias{newcomparePred}
\alias{newcovariate<-}
\alias{newDim}
\alias{newDim.default}
\alias{newDim.newpdNatural}
\alias{newDim.pdMat}
\alias{newDim.pdSymm}
\alias{newfdHess}
\alias{newfixed.effects}
\alias{newfixef}
\alias{newformula.reStruct}
\alias{newgetCovariate}
\alias{newgetCovariate.data.frame}
\alias{newgetCovariateFormula}
\alias{newgetData}
\alias{newgetFixDF}
\alias{newgetGroups}
\alias{newgetGroups.data.frame}
\alias{newgetGroups.lme}
\alias{newgetGroupsFormula}
\alias{newgetGroupsFormula.default}
\alias{newgetGroupsFormula.lme}
\alias{newgetGroupsFormula.reStruct}
\alias{newgetResponseFormula}
\alias{newInitialize}
\alias{newInitialize.reStruct}
\alias{newInitialize.rmincLmeStruct}
\alias{newintervals}
\alias{newisBalanced}
\alias{newisInitialized}
\alias{newisInitialized.pdMat}
\alias{newisInitialized.reStruct}
\alias{newlogDet}
\alias{newlogDet.reStruct}
\alias{newmatrix<-}
\alias{newmatrix<-.pdMat}
\alias{newmatrix<-.reStruct}
\alias{newMEdecomp}
\alias{newMEdims}
\alias{newMEEM}
\alias{newMEestimate}
\alias{newNames}
\alias{newNames.formula}
\alias{newNames.pdMat}
\alias{newNames.reStruct}
\alias{newNames<-}
\alias{newNames<-.pdMat}
\alias{newNames<-.reStruct}
\alias{newneedUpdate}
\alias{newneedUpdate.default}
\alias{newpdConstruct}
\alias{newpdConstruct.newpdLogChol}
\alias{newpdConstruct.newpdNatural}
\alias{newpdConstruct.pdMat}
\alias{newpdConstruct.pdSymm}
\alias{newpdFactor}
\alias{newpdFactor.newpdLogChol}
\alias{newpdFactor.newpdNatural}
\alias{newpdFactor.pdMat}
\alias{newpdFactor.pdSymm}
\alias{newpdFactor.reStruct}
\alias{newpdLogChol}
\alias{newpdMatrix}
\alias{newpdMatrix.newpdNatural}
\alias{newpdMatrix.pdMat}
\alias{newpdMatrix.pdSymm}
\alias{newpdMatrix.reStruct}
\alias{newpdNatural}
\alias{newrandom.effects}
\alias{newranef}
\alias{newrecalc}
\alias{newrecalc.modelStruct}
\alias{newrecalc.reStruct}
\alias{newsplitFormula}
\alias{newvarFunc}
\alias{newVariogram}
\alias{newvarWeights}
\alias{pdMat}
\alias{pdSymm}
\alias{plot.pdMat}
\alias{print.anatMatrix}
\alias{print.correlation}
\alias{print.mincMultiDim}
\alias{print.mincQvals}
\alias{print.mincSingleDim}
\alias{print.mincVoxel}
\alias{print.modelStruct}
\alias{print.pdMat}
\alias{print.reStruct}
\alias{print.rmincLme}
\alias{print.summary.lme}
\alias{print.summary.modelStruct}
\alias{print.summary.pdMat}
\alias{print.vertexMultiDim}
\alias{pt2}
\alias{residuals.rmincLmeStruct}
\alias{reStruct}
\alias{rmincLme}
\alias{rmincLme.formula}
\alias{rmincLmeApVar}
\alias{rmincLmeControl}
\alias{rmincLmeLoop}
\alias{rmincLmeScale}
\alias{rmincLmeStruct}
\alias{solve.newpdLogChol}
\alias{solve.newpdNatural}
\alias{solve.pdMat}
\alias{solve.pdSymm}
\alias{solve.reStruct}
\alias{summary.modelStruct}
\alias{summary.newpdLogChol}
\alias{summary.newpdNatural}
\alias{summary.pdMat}
\alias{summary.pdSymm}
\alias{summary.reStruct}
\alias{summary.rmincLme}
\alias{update.modelStruct}
\alias{update.reStruct}
\alias{update.rmincLme}
\alias{varWeights.rmincLmeStruct}
\alias{varWeights.varFunc}
\alias{vertexFDR}
\alias{vertexLm}
\alias{vertexTable}
\alias{wilcox.permutation}
\alias{wilcox.permutation.full}

\description{Internal RMINC objects.}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
\keyword{ internal }
