% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMINC.R
\docType{package}
\name{RMINC-package}
\alias{RMINC}
\alias{RMINC-package}
\title{RMINC: R interface to the MINC universe}
\description{
RMINC provides a means of accessing and performing statistical
analyses on MINC volumes. With facilities to calculate voxel-wise
statistics, perform FDR correction, rapidly visualize
results, and much more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Mouse-Imaging-Centre/RMINC}
  \item \url{https://wiki.mouseimaging.ca/display/MICePub/RMINC}
  \item Report bugs at \url{https://github.com/Mouse-Imaging-Centre/RMINC/issues}
}

}
\author{
\strong{Maintainer}: Chris Hammill \email{cfhammill@gmail.com}

Authors:
\itemize{
  \item Jason Lerch \email{jason.lerch@sickkids.ca}
  \item Matthijs van Eede
  \item Daniel Cassel
}

Other contributors:
\itemize{
  \item Jim Nikelski \email{nikelski@bic.mni.mcgill.ca} [contributor]
  \item Yohan Yee [contributor]
  \item Miriam Friedel [contributor]
  \item Nick Wang [contributor]
  \item Gabriel Devenyi \email{gdevenyi@gmail.com} [contributor]
  \item Benjamin Darwin [contributor]
  \item Bilal Syed [contributor]
}

}
