% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMINC.R
\name{RMINC_MASKED_VALUE}
\alias{RMINC_MASKED_VALUE}
\title{RMINC Masked Value}
\description{
RMINC_MASKED_VALUE is an object with the S3 class "RMINC_MASKED_VALUE" with NA value 
used to differentiate whether an NA was generated by the user or by an RMINC masking
procedure. The value, which will be written out with \link{mincWriteVol}, can be changed
to zero for example with \link{setRMINCMaskedValue}.
}

