% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{setRMINCMaskedValue}
\alias{setRMINCMaskedValue}
\title{Set Masked Value}
\usage{
setRMINCMaskedValue(val = NA)
}
\arguments{
\item{The}{new value for RMINC_MASKED_VALUE, defaults to resetting
the masked value to NA}
}
\description{
Use this function to configure how masked values should be treated
}

