% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{simplify_masked}
\alias{simplify_masked}
\title{Simplify Masked Results}
\usage{
simplify_masked(result_list)
}
\arguments{
\item{result_list}{The list of potentially masked results}
}
\value{
Scalar elements are converted to a vector,
Vector elements are converted to an array, data.frame elements are row-bound, 
and more complex objects are converted to a list. Result lists should be
all of the same type of unexpected errors may occur.
}
\description{
Take the results of an RMINC function that produces masked results
and coerce it to the appropriate result. Scalar elements are converted to a vector,
Vector elements are converted to an array, data.frame elements are row-bound, 
and more complex objects are converted to a list.
}

