% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_FDR.R
\name{thresholds}
\alias{thresholds}
\title{Get FDR Thresholds}
\usage{
thresholds(qvals)
}
\arguments{
\item{qvals}{A \code{mincQvals} object, typically computed with \code{mincFDR}
methods}
}
\value{
A matrix of thresholds, accessible with standard matrix indexing
}
\description{
Get FDR Thresholds
}

