% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{verboseRun}
\alias{verboseRun}
\title{Run function with/without output silenced}
\usage{
verboseRun(expr, verbose, env = parent.frame())
}
\arguments{
\item{expr}{an expression to run}

\item{verbose}{whether to permit output, or capture it}

\item{env}{the environment in which to run the expression, defaults
to the calling environment}
}
\description{
used in test bed
}

