% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vertex_statistics.R
\name{vertexApply}
\alias{vertexApply}
\title{Apply function over vertex Files}
\usage{
vertexApply(filenames, function.string)
}
\arguments{
\item{filenames}{vertex file names}

\item{function.string}{The function which to apply. Can only take a single
argument, which has to be 'x'.}
}
\value{
out: The output will be a single vector containing as many
         elements as there are vertices in the input variable.
}
\description{
This function is used to compute an arbitrary function of every region in a set of vertex files.
}
\examples{
\dontrun{
getRMINCTestData() 
gf = read.csv("/tmp/rminctestdata/CIVET_TEST.csv")
gf = civet.getAllFilenames(gf,"ID","TEST","/tmp/rminctestdata/CIVET","TRUE","1.1.12")
gf = civet.readAllCivetFiles("/tmp/rminctestdata/AAL.csv",gf)
vm <- vertexApply(gf$CIVETFILES$nativeRMStlink20mmleft,quote(mean(x)))
}
}

