% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsize.R
\name{vertexEffectSize}
\alias{vertexEffectSize}
\alias{mincEffectSize}
\alias{anatEffectSize}
\title{Effect Sizes}
\usage{
vertexEffectSize(buffer, predictors = NULL)

mincEffectSize(buffer, predictors = NULL)

anatEffectSize(buffer, predictors = NULL)
}
\arguments{
\item{buffer}{The results of a vertex/anat/mincLm run}

\item{predictors}{A vector of factor predictor names. By default the effect size
be computed for all treatment-coded factor columns.}
}
\value{
An object with columns of hedgesg-<factorlevel> and hedgesg_var-<factorlevel> for each
factor predictor in the GLM. The class and attributes of the input are preserved:
\itemize{
  \item \code{mincEffectSize}/\code{mincMultiDim} for voxel-wise inputs
  \item \code{vertexEffectSize}/\code{vertexMultiDim} for vertex-wise inputs
  \item \code{anatEffectSize}/\code{anatModel} for anatomy-wise inputs
}
Attributes \code{likeVolume}, \code{filenames}, \code{model}, \code{data}, \code{call},
\code{df}, \code{atlas}, \code{definitions}, and \code{stat-type} are carried over from
the input.
}
\description{
Takes the output of a minc modelling function and computes the unbiased
hedges g* and variance of hedges g*
}
\details{
This code implements the methods from Nakagawa, S., Cuthill, I.C., 2007. Effect size, confidence interval and statistical significance: a practical guide for biologists. Biol. Rev. Camb. Philos. Soc. 82, 591-05. https://doi.org/10.1111/j.1469-185X.2007.00027.x
for computing effect size of group comparisons from a GLM.

For now, interactions are explicitly excluded from being predictors. To get
effect size for interactions, use the interaction() function to create a new
treatment coded factor to use as a predictor.
}
\section{Functions}{
\itemize{
\item \code{mincEffectSize()}: mincEffectSize

\item \code{anatEffectSize()}: anatEffectSize

}}
\examples{
\dontrun{
getRMINCTestData()
# read the text file describing the dataset
gf <- read.csv("/tmp/rminctestdata/test_data_set.csv")
# run a linear model relating the data in all voxels to Genotype
vs <- mincLm(jacobians_fixed_2 ~ Sex, gf)
effectsize <- mincEffectSize(vs)
}
}
