% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_FDR.R
\name{vertexFDR}
\alias{vertexFDR}
\alias{vertexFDR.vertexMultiDim}
\alias{vertexFDR.vertexLmer}
\title{Vertex False Discovery Rates}
\usage{
vertexFDR(buffer, ...)

\method{vertexFDR}{vertexMultiDim}(buffer, ...)

\method{vertexFDR}{vertexLmer}(buffer, ...)
}
\arguments{
\item{buffer}{The results of a vertexLm type run.}

\item{...}{additional parameters to \link{mincFDR} like method}
}
\value{
A object of type \code{mincQvals} with the same number of columns
as the input. Each column now contains the qvalues for each vertex. Areas
outside the mask (if a mask was specified) will be represented by a
value of 1. The result also has an attribute called "thresholds"
which contains the 1, 5, 10, 15, and 20 percent false discovery rate
thresholds.
}
\description{
Takes the output of a minc modelling function and computes False Discovery Rate thresholds.
}
\section{Methods (by class)}{
\itemize{
\item \code{vertexMultiDim}: vertexMultiDim

\item \code{vertexLmer}: vertexLmer
}}

