% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_graphs.R
\name{vertexFindPeaks}
\alias{vertexFindPeaks}
\title{Vertex find peaks}
\usage{
vertexFindPeaks(
  data_map,
  graph,
  mindist = 1,
  direction = c("both", "positive", "negative"),
  threshold = 0,
  output = c("mask", "indices")
)
}
\arguments{
\item{data_map}{A vector or text file of values to search for peaks}

\item{graph}{An igraph object describing the connectivity, in most cases this will be
a mesh graph produced by \link{obj_to_graph}}

\item{mindist}{the search radius (in number of neighbours) to consider when finding a peak. A
node is considered a peak if it is more extreme than any element in it's neighbourhood up to
\code{mindist} edges away}

\item{direction}{Either positive, negative, or both indicating whether to consider maxima,
minima or both.}

\item{threshold}{threshold above or below which to consider peaks. In the both direction case
a two element vector may be passed indicating positive and negative thresholds respectively. 
If a single element is passed in the both direction case, it is treated as \code{c(threshold, -threshold)}.
Defaults to 0.}

\item{output}{Either mask or indices, indicating whether a logical mask or a vector of indices is desired.}
}
\value{
Either a logical vector indicating peaks or a vector of peak indices determined by \code{output}
}
\description{
Find extrema of set of values given a connectivity graph
}
\details{
Note that if coordinates of peaks are desired, they can be accessed from the parent \code{bic_obj}
by \code{obj$vertex_matrix[,peaks]}
}
