% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vertex_statistics.R
\name{vertexLmer}
\alias{vertexLmer}
\title{Vertex Mixed Effects Models}
\usage{
vertexLmer(formula, data, mask = NULL, parallel = NULL, REML = TRUE,
  control = lmerControl(), start = NULL, verbose = 0L,
  temp_dir = getwd(), safely = FALSE, cleanup = TRUE)
}
\arguments{
\item{formula}{the lmer formula, filenames go on left hand side}

\item{data}{the data frame, all items in formula should be in here}

\item{mask}{the mask within which lmer is solved}

\item{parallel}{how many processors to run on (default=single processor).
Specified as a two element vector, with the first element corresponding to
the type of parallelization, and the second to the number
of processors to use. For local running set the first element to "local" or "snowfall"
for back-compatibility, anything else will be run with BatchJobs see \link{pMincApply}
and \link{configureMincParallel} for details.
Leaving this argument NULL runs sequentially and may take a long time.}

\item{REML}{whether to use use Restricted Maximum Likelihood or Maximum Likelihood}

\item{control}{lmer control function}

\item{start}{lmer start function}

\item{verbose}{lmer verbosity control}

\item{temp_dir}{A directory to create temporary mask and registry files if 
\code{parallel = c("sge", n)}. This should not be \code{tempdir()} so that workers can see
these files. Defaults to the current working directory.}

\item{safely}{whether or not to wrap the per-voxel lmer code in an exception catching
block (\code{tryCatch}), when TRUE this will downgrade errors to warnings and return
NA for the result.}

\item{cleanup}{Whether or not to cleanup registry files after a queue parallelized 
run}
}
\description{
Perform linear mixed effects model fitting for vertex data.
vertexLmer should be used the same way as a straight lmer call, except
that the left hand side of the equation contains vertex filenames rather than
an actual response variable.
}
\details{
\code{vertexLmer}, like its relative \link{mincLmer} provides an interface to running 
linear mixed effects models at every vertex. Unlike standard linear models testing hypotheses 
in linear mixed effects models is more difficult, since the denominator degrees of freedom are 
more difficult to  determine. RMINC provides two alternatives: (1) estimating degrees of freedom using the
\code{\link{mincLmerEstimateDF}} function, and (2) comparing two separate models using
\code{\link{mincLogLikRatio}} (which in turn can be corrected using
\code{\link{mincLogLikRatioParametricBootstrap}}). For the most likely models - longitudinal
models with a separate intercept or separate intercept and slope per subject - both of these
approximations are likely correct. Be careful in using these approximations if
using more complicated random effects structures.
}
\seealso{
\code{\link{lmer}} for description of lmer and lmer formulas; \code{\link{mincLm}}
}

