% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{vertexTable}
\alias{vertexTable}
\title{Create a table of vertex values}
\usage{
vertexTable(filenames)
}
\arguments{
\item{filenames}{paths to the vertex data files}
}
\value{
a matrix where each `column` is a matrix of vertex
data corresponding to a single file
}
\description{
Read files containing vertex data into a matrix
}
