% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vertex_statistics.R
\name{vertexTFCE}
\alias{vertexTFCE}
\alias{vertexTFCE.numeric}
\alias{vertexTFCE.matrix}
\title{Threshold Free Cluster Enhancement}
\usage{
vertexTFCE(x, ...)

\method{vertexTFCE}{numeric}(x, surface, d = 0.1, E = 0.5, H = 2,
  nsteps = 100, side = c("both", "positive", "negative"), weights = NULL,
  ...)

\method{vertexTFCE}{matrix}(x, surface, E = 0.5, H = 2, nsteps = 100,
  side = c("both", "positive", "negative"), weights = NULL, ...)
}
\arguments{
\item{x}{A numeric vector, a filepath to a set of values,
or a \code{matrix} object, or \code{vertexLm} object.}

\item{...}{additional arguments for methods}

\item{surface}{Either a mesh object corresponding to the surface,
an igraph graph object of surface created by \link{obj_to_graph}, or an adjacency list (see details). 
For the \code{matrix} and \link{vertexLm} cases, either a single surface object may be passed and
used for each individual, or a vector of file names}

\item{d}{The discretization step-size for approximating the threshold integral (default .1)}

\item{E}{The exponent by which to raise the extent statistic (default .5)}

\item{H}{The exponent by which to raise the height (default 2)}

\item{nsteps}{The number of steps to discretize the TFCE computation over}

\item{side}{Whether to consider positive and negative statistics or both (default both)}

\item{weights}{A weighting vector assigning area to vertices. The default varies by 
\code{surface} object type. If \code{surface} is an \code{bic_obj} the area of each
vertex is equal to 1/3 the sum of the areas of the triangles it is a member of. If \code{surface}
is an \code{igraph} or adjacency list the default is a vector of ones.}
}
\value{
The behaviour of \code{vertexTFCE} is to perform cluster free enhancement on a object,
in the single dimensional case, a string denoting a vertex file or a numeric vector
it returns a numeric vector with the result. In the matrix case each column is cluster enhanced and 
recomposed into a matrix.
In the \link{vertexLm} case a randomization test is performed with the t-stats enhanced computing the maximal
value after enhancement. A vector of maxima is returned to compute exceedence probabilities, along with
the original statistic after cluster enhancement.
}
\description{
Perform threshold free cluster enhancement as described in 
Smith and Nichols (2008). Cluster-like structures are enhanced
to allow a hybrid cluster/voxel analysis to be performed.
}
\details{
Passing an adjacency list will save some compute time
but is not recommended for general use. If an adjacency list is passed should index starting from 0 for compatibility with c++
code. Adjacency lists of this kind can be generated from graphs with \code{lapply(as_adj_list(graph), `-`, 1)}
using the \link{as_adj_list} from the igraph library.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: numeric

\item \code{matrix}: matrix
}}

