% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{writeVertex}
\alias{writeVertex}
\title{Writes vertex data to a file with an optional header}
\usage{
writeVertex(vertexData, filename, headers = TRUE, mean.stats = NULL,
  gf = NULL)
}
\arguments{
\item{vertexData}{vertex data to be written}

\item{filename}{full path to file where data shall be written}

\item{headers}{Whether or not to write header information}

\item{mean.stats}{mean vertex data that may also be written}

\item{gf}{glim matrix that can be written to the header}
}
\value{
A file is generated with the vertex data and optional headers
}
\description{
Writes vertex data to a file with an optional header
}
\examples{
getRMINCTestData() 
gf = read.csv("/tmp/rminctestdata/CIVET_TEST.csv")
gf = civet.getAllFilenames(gf,"ID","TEST","/tmp/rminctestdata/CIVET","TRUE","1.1.12")
gf = civet.readAllCivetFiles("/tmp/rminctestdata/AAL.csv",gf)
writeVertex(gf$nativeRMStlink20mm,"~/RMStlink20mm.txt",FALSE,NULL,NULL)
}

