% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vis3D.R
\name{add_colour_bar}
\alias{add_colour_bar}
\title{Add a colour bar for a mesh}
\usage{
add_colour_bar(mesh, title = "", ..., column_widths = c(0.75, 0.25),
  which_col = length(column_widths))
}
\arguments{
\item{mesh}{A \code{obj_mesh} object created with \link{colour_mesh}}

\item{title}{The label to give the colour bar}

\item{...}{additional plot parameters to be passed down to the \link{image}
function.}

\item{column_widths}{define the horizonatal segmentation of the plot device,
values are scaled to sum to 1. Defaults to a 15% region on the left to hold
the colour bar}

\item{which_col}{The column index of where to draw the colour bar, defaults
to the last panel, useful for creating a non-standard colour bar location.}
}
\value{
invisible NULL
}
\description{
Add a colour bar that corresponds to the colours in a given colourized mesh
to the current rgl device
}
