% Generated by roxygen2 (4.0.1): do not edit by hand
\name{anatAnova}
\alias{anatAnova}
\title{Performs ANOVA on each region specified}
\usage{
anatAnova(formula, data = NULL, anat = NULL, subset = NULL)
}
\arguments{
\item{formula}{a model formula}

\item{data}{a data.frame containing variables in formula}

\item{ant}{an array of atlas labels vs subject data}

\item{subset}{rows to be used, by default all are used}
}
\value{
Returns an array with the F-statistic for each model specified by formula with the following attributes: model – design matrix, filenames –
	, stat-type: type of statistic used, df – degrees of freedom of each statistic.
}
\description{
Performs ANOVA on each region specified
}
\examples{
getRMINCTestData()
gf = read.csv("/tmp/rminctestdata/CIVET_TEST.csv")
gf = civet.getAllFilenames(gf,"ID","TEST","/tmp/rminctestdata/CIVET","TRUE","1.1.12")
gf = civet.readAllCivetFiles("/tmp/rminctestdata/AAL.csv",gf)
rmincAnova = anatAnova(~ Sex,gf,gf$lobeThickness);
}
\seealso{
mincAnova,vertexAnova
}

