% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_anatomy.R
\name{anatGetFile2}
\alias{anatGetFile2}
\title{Get another file}
\usage{
anatGetFile2(filenames, atlas, method = c("jacobians", "labels", "sums",
  "means"), mask = NULL, strict = TRUE)
}
\arguments{
\item{filenames}{... asdf}

\item{atlas}{... asdf}

\item{method}{... asdf}

\item{mask}{... asdf}

\item{strict}{... asdf}
}
\value{
something
}
\description{
doco
}

