% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_funcs.R
\name{anatLm}
\alias{anatLm}
\title{Calculates statistics and coefficients for linear model of specified anat structure}
\usage{
anatLm(formula, data, anat, subset = NULL)
}
\arguments{
\item{formula}{a model formula. The RHS of the formula may contain one term with a matrix. If
so only the + operator may be used, and only two terms may appear on the RHS}

\item{data}{a data.frame containing variables in formula}

\item{anat}{an array of atlas labels vs subject data}

\item{subset}{rows to be used, by default all are used}
}
\value{
Returns an object containing the R-Squared,value,coefficients,F 
and t statistcs that can be passed directly into anatFDR. Additionally
has the attributes for model,stat type and degrees of freedom.
}
\description{
Calculates statistics and coefficients for linear model of specified anat structure
}
\examples{
getRMINCTestData() 
gf = read.csv("/tmp/rminctestdata/CIVET_TEST.csv")
gf = civet.getAllFilenames(gf,"ID","TEST","/tmp/rminctestdata/CIVET","TRUE","1.1.12")
gf = civet.readAllCivetFiles("/tmp/rminctestdata/AAL.csv",gf)
rmincLm= anatLm(~ Sex,gf,gf$lobeThickness); 
}
\seealso{
mincLm,anatLm,anatFDR
}

