% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_anatomy.R
\name{anatLmer}
\alias{anatLmer}
\title{Anatomy Linear Mixed Effects Model}
\usage{
anatLmer(formula, data, anat, subset = NULL, REML = TRUE,
  control = lmerControl(), verbose = FALSE, start = NULL)
}
\arguments{
\item{formula}{the model formula}

\item{data}{the predictor variables}

\item{anat}{a subject by label matrix of anatomical 
summaries typically produced by \link{anatGetAll}}

\item{subset}{rows to subset}
}
\value{
an \code{anatModel} object of statistical results
}
\description{
Apply a linear mixed effects model to the results of
\link{anatGetAll}.
}

