% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_anatomy.R
\name{anatLmer}
\alias{anatLmer}
\title{Anatomy Linear Mixed Effects Model}
\usage{
anatLmer(formula, data, anat, REML = TRUE, control = lmerControl(),
  verbose = FALSE, start = NULL, parallel = NULL, safely = FALSE,
  summary_type = "fixef", weights = NULL, resources = list(),
  conf_file = getOption("RMINC_BATCH_CONF"))
}
\arguments{
\item{formula}{the lmer formula, filenames go on left hand side}

\item{data}{the data frame, all items in formula should be in here}

\item{anat}{a subject by label matrix of anatomical 
summaries typically produced by \link{anatGetAll}}

\item{REML}{whether to use use Restricted Maximum Likelihood or Maximum Likelihood}

\item{control}{lmer control function}

\item{verbose}{lmer verbosity control}

\item{start}{lmer start function}

\item{parallel}{how many processors to run on (default=single processor).
Specified as a two element vector, with the first element corresponding to
the type of parallelization, and the second to the number
of processors to use. For local running set the first element to "local" or "snowfall"
for back-compatibility, anything else will be run with batchtools see \link{pMincApply}.
Leaving this argument NULL runs sequentially and may take a long time.}

\item{safely}{whether or not to wrap the per-voxel lmer code in an exception catching
block (\code{tryCatch}), when TRUE this will downgrade errors to warnings and return
NA for the result.}

\item{summary_type}{Either one of
\itemize{
  \item{fixef: default and equivalent to older versions of RMINC, returns fixed effect coefficients and t-values}
  \item{ranef: returns random effect coefficients and t-values}
  \item{both: both fixed and random effects}
  \item{anova: return the F-statistic for each fixed effect}
} or a function to be used to generate the summary}

\item{weights}{weights to be applied to each observation}

\item{resources}{A list of resources to use for the jobs, for example
\code{ list(nodes = 1, memory = "8G", walltime = "01:00:00") }. See
\code{system.file("parallel/pbs_script.tmpl", package = "RMINC")} and
\code{system.file("parallel/sge_script.tmpl", package = "RMINC")} for
more examples}

\item{conf_file}{A batchtools configuration file}
}
\description{
Fit a linear mixed effects model for each structure in the results of
\link{anatGetAll}.
}
\details{
\code{anatLmer}, like its relative \link{mincLmer} provides an interface to running 
linear mixed effects models at every vertex. Unlike standard linear models testing hypotheses 
in linear mixed effects models is more difficult, since the denominator degrees of freedom are 
more difficult to  determine. RMINC provides estimating degrees of freedom using the
\code{\link{anatLmerEstimateDF}} function. For the most likely models - longitudinal
models with a separate intercept or separate intercept and slope per subject - this
approximation is likely correct. Be careful in using this approximations if
using more complicated random effects structures.
}
