% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_anatomy.R
\name{anatLmer}
\alias{anatLmer}
\title{Anatomy Linear Mixed Effects Model}
\usage{
anatLmer(formula, data, anat, REML = TRUE, control = lmerControl(),
  verbose = FALSE, start = NULL, parallel = NULL, safely = FALSE)
}
\arguments{
\item{formula}{the lmer formula, filenames go on left hand side}

\item{data}{the data frame, all items in formula should be in here}

\item{anat}{a subject by label matrix of anatomical 
summaries typically produced by \link{anatGetAll}}

\item{REML}{whether to use use Restricted Maximum Likelihood or Maximum Likelihood}

\item{control}{lmer control function}

\item{verbose}{lmer verbosity control}

\item{start}{lmer start function}

\item{parallel}{how many processors to run on (default=single processor).
Specified as a two element vector, with the first element corresponding to
the type of parallelization, and the second to the number
of processors to use. For local running set the first element to "local" or "snowfall"
for back-compatibility, anything else will be run with BatchJobs see \link{pMincApply}
and \link{configureMincParallel} for details.
Leaving this argument NULL runs sequentially and may take a long time.}

\item{safely}{whether or not to wrap the per-voxel lmer code in an exception catching
block (\code{tryCatch}), when TRUE this will downgrade errors to warnings and return
NA for the result.}
}
\description{
Fit a linear mixed effects model for each structure in the results of
\link{anatGetAll}.
}
\details{
\code{anatLmer}, like its relative \link{mincLmer} provides an interface to running 
linear mixed effects models at every vertex. Unlike standard linear models testing hypotheses 
in linear mixed effects models is more difficult, since the denominator degrees of freedom are 
more difficult to  determine. RMINC provides two alternatives: (1) estimating degrees of freedom using the
\code{\link{mincLmerEstimateDF}} function, and (2) comparing two separate models using
\code{\link{mincLogLikRatio}} (which in turn can be corrected using
\code{\link{mincLogLikRatioParametricBootstrap}}). For the most likely models - longitudinal
models with a separate intercept or separate intercept and slope per subject - both of these
approximations are likely correct. Be careful in using these approximations if
using more complicated random effects structures.
}

