% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_funcs.R
\name{anatSummaries}
\alias{anatMean}
\alias{anatSd}
\alias{anatSum}
\alias{anatSummaries}
\alias{anatVar}
\alias{vertexMean}
\alias{vertexSd}
\alias{vertexSum}
\alias{vertexVar}
\title{Create descriptive statistics across an anat structure}
\usage{
anatMean(anat)
}
\arguments{
\item{anat}{anat structure.}
}
\value{
out: The output will be a single vector containing as many
         elements as there are regions in the input variable.
}
\description{
This function is used to compute the mean, standard deviation,
   		sum, or variance of every region in an anat structure.
}
\examples{
getRMINCTestData() 
gf = read.csv("/tmp/rminctestdata/CIVET_TEST.csv")
gf = civet.getAllFilenames(gf,"ID","TEST","/tmp/rminctestdata/CIVET","TRUE","1.1.12")
gf = civet.readAllCivetFiles("/tmp/rminctestdata/AAL.csv",gf)
vm <- anatMean(gf$lobeThickness)
}

