% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_funcs.R
\name{anatSummaries}
\alias{anatMean}
\alias{anatSd}
\alias{anatSum}
\alias{anatSummaries}
\alias{anatVar}
\title{anatSummaries}
\usage{
anatMean(anat)

anatSum(anat)

anatVar(anat)

anatSd(anat)
}
\arguments{
\item{anat}{anat structure.}
}
\value{
out: The output will be a single vector containing as many
         elements as there are regions in the input variable.
}
\description{
These functions are used to compute the mean, standard deviation,
   		sum, or variance of every region in an anat structure.
}
\section{Functions}{
\itemize{
\item \code{anatMean}: mean

\item \code{anatSum}: sum

\item \code{anatVar}: variance

\item \code{anatSd}: standard deviation
}}
\examples{
\dontrun{
getRMINCTestData() 
gf = read.csv("/tmp/rminctestdata/CIVET_TEST.csv")
gf = civet.getAllFilenames(gf,"ID","TEST","/tmp/rminctestdata/CIVET","TRUE","1.1.12")
gf = civet.readAllCivetFiles("/tmp/rminctestdata/AAL.csv",gf)
vm <- anatMean(gf$lobeThickness)
}
}

