% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_anatomy.R
\name{as.data.frame.anatMatrix}
\alias{as.data.frame.anatMatrix}
\title{Create Anatomy data.frame}
\usage{
\method{as.data.frame}{anatMatrix}(x, ...)
}
\arguments{
\item{x}{An \code{anatMatrix} object produced by \link{anatGetAll}}

\item{...}{additional parameters for methods}
}
\value{
A data.frame augmented with addition attributes from anat
}
\description{
Convert an anatomy frame to data.frame for use with tidy tools
}
\details{
Note that structure names are munged such that each structure name has non-standard
characters replaced by underscores
}

