% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civet.R
\name{civet.checkVersion}
\alias{civet.checkVersion}
\title{Check for a known Civet Version}
\usage{
civet.checkVersion(civetVersion)
}
\arguments{
\item{civetVersion}{A string specifying the Civet version number, e.g.,
\dQuote{1.1.7}}
}
\value{
Nothing is returned.
}
\description{
As different versions of Civet may reflect changes in filename or directory
structure, it is important that the civet.* functions be validated for each
new version of Civet.  This function checks whether a given Civet version
number has been validated for use by these routines.
}
\details{
If the passed Civet version cannot be validated, a warning message is sent
to std output.
}
\examples{

\dontrun{
civetVersion <- "1.1.8"
civet.checkVersion(civetVersion)
}
}
\author{
Jim Nikelski \email{nikelski@bic.mni.mcgill.ca}
}
