% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civet.R
\name{civet.getAllFilenames}
\alias{civet.getAllFilenames}
\title{civet.getAllFilenames}
\usage{
civet.getAllFilenames(
  gf,
  idvar,
  prefix,
  basedir,
  append = TRUE,
  civetVersion = "1.1.9",
  cnf = NULL
)
}
\arguments{
\item{gf}{Data Frame with subject information}

\item{idvar}{column name in gf with subject IDs}

\item{prefix}{Prefix specified when CIVET was run}

\item{basedir}{directory where all CIVET output was stored}

\item{append}{Whether to append the results to the input gf}

\item{civetVersion}{Version of CIVET}

\item{cnf}{A list of configuration information used to parse CIVET files produced by CBRAIN. Unnecessary for
CIVETs < 2. Should be read from CBRAIN yaml config file, but can be input manually as a list containing
thickness_method (tlink or laplacian), thickness_kernel (in mm), and atlas as either AAL or DKT}
}
\value{
gf is returned with CIVET filenames
}
\description{
Generates list of filenames output by CIVET
}
\details{
Prior to running, read.csv  may be called to generate the input argument gf. 
The results will be stored under the column name CIVETFILES either in the input gf (if append = TRUE) or in a new gf. 
Currently only CIVET versions 1.1.9 and 1.1.12 are supported.
}
\examples{
\dontrun{
getRMINCTestData() 
gf = read.csv("/tmp/rminctestdata/CIVET_TEST.csv")
gf = civet.getAllFilenames(gf,"ID","TEST","/tmp/rminctestdata/CIVET","TRUE","1.1.12")
}
}
\seealso{
civet.readAllCivetFiles
}
