% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civet.R
\name{civet.getFilename}
\alias{civet.getFilename}
\alias{civet.getFilenameCerebrumMask}
\alias{civet.getFilenameClassify}
\alias{civet.getFilenameCorticalThickness}
\alias{civet.getFilenameCsfPve}
\alias{civet.getFilenameGrayMatterPve}
\alias{civet.getFilenameGrayMatterSurfaces}
\alias{civet.getFilenameLinearTransform}
\alias{civet.getFilenameMeanSurfaceCurvature}
\alias{civet.getFilenameMidSurfaces}
\alias{civet.getFilenameNonlinearTransform}
\alias{civet.getFilenameSkullMask}
\alias{civet.getFilenameStxT1}
\alias{civet.getFilenameWhiteMatterPve}
\alias{civet.getFilenameWhiteMatterSurfaces}
\title{Get Selected Civet Filenames}
\usage{
civet.getFilenameClassify(scanID, baseDir, civetVersion = "1.1.9",
  fullPath = TRUE)

civet.getFilenameGrayMatterPve(scanID, baseDir, civetVersion = "1.1.9",
  fullPath = TRUE)

civet.getFilenameWhiteMatterPve(scanID, baseDir, civetVersion = "1.1.9",
  fullPath = TRUE)

civet.getFilenameCsfPve(scanID, baseDir, civetVersion = "1.1.9",
  fullPath = TRUE)

civet.getFilenameStxT1(scanID, baseDir, civetVersion = "1.1.9",
  fullPath = TRUE)

civet.getFilenameCerebrumMask(scanID, baseDir, civetVersion = "1.1.9",
  fullPath = TRUE)

civet.getFilenameSkullMask(scanID, baseDir, civetVersion = "1.1.9",
  fullPath = TRUE)

civet.getFilenameGrayMatterSurfaces(scanID, baseDir, civetVersion = "1.1.9",
  fullPath = TRUE)

civet.getFilenameWhiteMatterSurfaces(scanID, baseDir, civetVersion = "1.1.9",
  fullPath = TRUE)

civet.getFilenameMidSurfaces(scanID, baseDir, civetVersion = "1.1.9",
  fullPath = TRUE)

civet.getFilenameCorticalThickness(scanID, baseDir, civetVersion = "1.1.9",
  fullPath = TRUE)

civet.getFilenameMeanSurfaceCurvature(scanID, baseDir, civetVersion = "1.1.9",
  fullPath = TRUE)

civet.getFilenameLinearTransform(scanID, baseDir, civetVersion = "1.1.9",
  fullPath = TRUE)

civet.getFilenameNonlinearTransform(scanID, baseDir, civetVersion = "1.1.9",
  fullPath = TRUE)
}
\arguments{
\item{scanID}{A string specifying the unique scan-id (and thus
sub-directory) within the Civet root output directory.}

\item{baseDir}{A string specifying the Civet root output directory.  This
directory will, in turn, contain all of the scanIDs.}

\item{civetVersion}{An optional string specifying the version of Civet used
to create the output.  This is significant since filenames and directory
structures may change across difference versions of Civet.}

\item{fullPath}{A boolean specifying whether the function is to return
either a fully-qualified path (TRUE) or just the filename without path
(FALSE).}
}
\value{
Either a string or a list is returned, depending on the number of
filenames returned.  Specifically, a single filename is returned as a
string, whereas multiple filenames are returned as named lists.
}
\description{
Returns either one or more Civet filenames, depending on file type.
}
\details{
The purpose of this function is to facilitate writing code requiring
manipulation of Civet products.  To this purpose, we have written a number
of convenience functions which, given the type of file desired and a path to
the Civet output directory, are able to determine and return the actual
filename(s).
}
\section{Functions}{
\itemize{
\item \code{civet.getFilenameClassify}: Tissue classification

\item \code{civet.getFilenameGrayMatterPve}: gray matter pve

\item \code{civet.getFilenameWhiteMatterPve}: white matter pve

\item \code{civet.getFilenameCsfPve}: csf pve

\item \code{civet.getFilenameStxT1}: Standard to T1 transform

\item \code{civet.getFilenameCerebrumMask}: brain mask

\item \code{civet.getFilenameSkullMask}: skull mask

\item \code{civet.getFilenameGrayMatterSurfaces}: gray matter surfaces

\item \code{civet.getFilenameWhiteMatterSurfaces}: white matter surfaces

\item \code{civet.getFilenameMidSurfaces}: mid surfaces

\item \code{civet.getFilenameCorticalThickness}: cortical thickness

\item \code{civet.getFilenameMeanSurfaceCurvature}: surface curvature

\item \code{civet.getFilenameLinearTransform}: linear transform

\item \code{civet.getFilenameNonlinearTransform}: non-linear transform
}}
\examples{

\dontrun{
library(RMINC)

# set Civet root path and scan-identifier
basePath <- "~/tmp/ADNI/civet/pipeOut"
scanID = "0221-M-AD"

# get the name of the aggregate tissue classification volume
# ... and then read it
classifyVolname <- civet.getFilenameClassify(scanID, basePath)
classifyVol <- mincIO.readVolume(classifyVolname)

# get the left and right gray matter surface filenames and then 
# ... print the names
gmSurfName <- civet.getFilenameGrayMatterSurfaces(scanID, basePath)
print(gmSurfName$left)
print(gmSurfName$right)

# get the various transformation file filenames
lin.xfmName <- civet.getFilenameLinearTransform(scanID, basePath)
print(lin.xfmName)
nlin.xfmNames <- civet.getFilenameNonlinearTransform(scanID, basePath)
print(nlin.xfmNames$xfm)		# name of the nlin xfm file
print(nlin.xfmNames$grid)		# name of the nlin grid file
}
}
\author{
Jim Nikelski \email{nikelski@bic.mni.mcgill.ca}
}

