% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{getRMINCTestData}
\alias{getRMINCTestData}
\title{Download Example Data}
\usage{
getRMINCTestData(
  dataPath = getOption("RMINC_DATA_DIR", tempdir()),
  method = "libcurl"
)
}
\arguments{
\item{dataPath}{The directory to download and unpack the test data 
(unpacks in dataPath/rminctestdata). Default can be set with the option RMINC_DATA_DIR which
can in turn be set with the environment variable RMINC_DATA_DIR. If unset a temporary directory
is created.}

\item{method}{Argument to pass to \link{download.file} typical options are \code{libcurl}
and \code{wget}}
}
\description{
Download the example data needed to run our examples in your /tmp directory
The data can be downloaded manually from 
\url{https://wiki.mouseimaging.ca/download/attachments/1654/rminctestdata.tar.gz}
}
