% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyRMINC.R
\name{launch_shinyRMINC}
\alias{launch_shinyRMINC}
\title{launch a shiny based inspector}
\usage{
launch_shinyRMINC(statsoutput, anatVol, volumes = NULL, keepBetas = FALSE,
  plotcolumns = NULL, modelfunc = NULL)
}
\arguments{
\item{statsoutput}{the output of mincLm or mincAnova}

\item{anatVol}{the anatomical volume on which to display the statistics. Can
be passed a filename, a MINC volume (from mincGetVol), or a mincArray.}

\item{volumes}{a matrix or data frame of volumes for plotting}

\item{keepBetas}{whether to include the beta coefficients}

\item{plotcolumns}{extra data to be used for plotting}

\item{modelfunc}{optional modelling function}
}
\description{
This uses shiny to inspect the output of mincLm or mincAnova. Produces
various views, the ability to plot individual voxels, and get the output of
the stats model at that voxel.
}
\examples{
\dontrun{
vs <- mincLm(reljacobians02 ~ sex*treatment, subset(gfs, treatment != "None"))
anatVol <- mincArray(mincGetVolume("anatomyfile.mnc"))
launch_shinyRMINC(vs, anatVol, volumes=gfs$vols, plotcolumns=gfs[,c("sex", "Neonatal")], keepBetas=F)
}
}

