% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_parallel.R
\name{mcMincApply}
\alias{mcMincApply}
\title{Local multicore mincApply}
\usage{
mcMincApply(filenames, fun, ..., mask = NULL, tinyMask = FALSE,
  batches = 4, temp_dir = tempdir(), cores = getOption("mc.cores",
  parallel::detectCores() - 1), return_raw = FALSE, cleanup = TRUE,
  collate = simplify2minc)
}
\arguments{
\item{filenames}{Paths to the minc files to apply accross}

\item{fun}{An arbitrary R function to be applied}

\item{...}{Additional arguments to pass to fun, see details for a warning}

\item{mask}{The mask used to select voxels to apply to}

\item{tinyMask}{Shrink the mask for testing}

\item{batches}{The number of batches to divide the job into}

\item{temp_dir}{A directory to hold mask files used in the job batching}

\item{cores}{the Number of cores to use, defaults to the option
\code{mc.cores} or 2 if it is unset}

\item{return_raw}{An internal use argument that prevents the resulting object
from being reordered and expanded.}

\item{collate}{A function to collate the list into another object type}
}
\description{
Split a minc apply job into batches and process it locally
using a fork cluster generated by the parallel package.
}

