\name{minc.dimensions.sizes}
\alias{minc.dimensions.sizes}

\title{Return dimension sizes of MINC2 volume}

\description{Returns the dimension sizes of a MINC2 volume}

\usage{
minc.dimensions.sizes(filename)
}

\arguments{
  \item{filename}{The filename of the MINC2 volume whose dimension sizes
    are to be returned.}
}

\seealso{minc.get.volume}
