% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mincApply}
\alias{mincApply}
\alias{pMincApply}
\title{Apply arbitrary R function at every voxel}
\usage{
mincApply(filenames, quote(mean(x)), mask="mask.mnc")
pMincApply(filenames, quote(mean(x)), mask="mask.mnc", workers=4,
           method="snowfall", tinyMask=FALSE, global = "",packages = "")
}
\arguments{
\item{filenames}{The MINC files over which to apply the function. Have to be
the same sampling.}

\item{function-string}{The function which to apply. Can only take a single
argument, which has to be 'x'. See details and examples.}

\item{mask}{The filename of a mask - function will only be evaluated
inside the mask.}

\item{maskval}{An integer specifying the value inside the mask where to
apply the function. If left blank (the default) then anything
above 0.5 will be considered inside the mask. This argument
only works for mincApply, not pMincApply.}

\item{workers}{The number of processes to split the function into. Only
works for pMincApply, not for mincApply.}

\item{method}{The method to be used for parallization ("snowfall" or "sge")}

\item{tinyMask}{Only for pMincApply; if set to some numeric value it computes
the function over that fraction of the mask. Useful for
debugging purposes only (i.e. if you want to test out whether
a new function works across the cluster.)}

\item{global}{Global variables to be exported}

\item{packages}{R Packages to be exported}
}
\value{
out The output is a matrix with the same number of rows a the
file sizes and the same number of columns as output by the
function that was applied. Cast into one of the MINC classes
to make writing it out with mincWriteVolume easier.
}
\description{
Can execute any R function at every voxel for a set of MINC files
}
\details{
mincApply allows one to execute any R function at every voxel of a
set of files. There are two variants: mincApply, which works
inside the current R session, and pMincApply, which uses the
snowfall and Rsge packages to split the execution of
the function across multiple cores/processors on the same machine
or across a cluster.
Unless the function to be applied takes a single argument a
wrapper function has to be written. This is illustrated in the
examples; briefly, the wrapper function takes a single argument,
called 'x', which will take on the voxel values at every voxel.
The function has to be turned into a string; the quote function
can be very handy. The output of this function also has to be a
simple vector or scalar.
Note that interpreted R can be very slow. Mindnumbingly slow. It
therefore pays to write optimal functions or, whenever available,
use the optimized equivalents. In other words and to give one
example, use mincLm rather than applying lm, and if lm has to
really be applied, try to use lm.fit rather than plain lm.
}
\examples{
getRMINCTestData()
gf <- read.csv("/tmp/rminctestdata/test_data_set.csv")
ma <- mincApply(gf$jacobians_fixed_2,quote(mean(x)));
mincWriteVolume(ma, "means.mnc")
### run the whole thing in parallel on the local machine
testFunc <- function (x) { return(c(1,2))}
pout <- pMincApply(gf$jacobians_fixed_2, quote(testFunc(x)),workers = 4,global = c('gf','testFunc'))
mincWriteVolume(pout, "pmincOut.mnc")
}
\seealso{
mincWriteVolume,mincMean,mincSd,mincLm,mincAnova
}

