% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{mincConvertWorldToVoxel}
\alias{mincConvertWorldToVoxel}
\title{World to Voxel}
\usage{
mincConvertWorldToVoxel(filename, v1, v2, v3, nearest_voxel = TRUE)
}
\arguments{
\item{filename}{A path to a minc file}

\item{v1}{First world coordinate}

\item{v2}{Second world coordinate}

\item{v3}{Third world coordinate}

\item{nearest_voxel}{Logical whether to round the result to the nearest voxel}
}
\value{
a 3-component numeric vector of voxel coordinates
}
\description{
Convert coordinates in the volumes world space (a continuous coordinate space)
to a discrete set of voxel coordinates
}

