% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_FDR.R
\name{mincFDR}
\alias{mincFDR}
\alias{mincFDR.mincLmer}
\alias{mincFDR.mincLogLikRatio}
\alias{mincFDR.mincMultiDim}
\alias{mincFDR.mincSingleDim}
\title{False Discovery Rates}
\usage{
mincFDR(buffer, ...)

\method{mincFDR}{mincSingleDim}(buffer, df, mask = NULL, method = "qvalue",
  ...)

\method{mincFDR}{mincLogLikRatio}(buffer, mask = NULL, ...)

\method{mincFDR}{mincLmer}(buffer, mask = NULL, ...)

\method{mincFDR}{mincMultiDim}(buffer, columns = NULL, mask = NULL,
  df = NULL, method = "FDR", statType = NULL, ...)
}
\arguments{
\item{buffer}{The results of a mincLm type run.}

\item{...}{extra parameters to pass to methods}

\item{df}{The degrees of freedom - normally this can be determined
from the input object.}

\item{mask}{Either a filename or a numeric vector representing a mask
only values inside the mask will be used to compute the
threshold.}

\item{method}{The method used to compute the false discovery
rate. Options are "FDR" and "pFDR".}

\item{columns}{A vector of column names. By default the threshold will
be computed for all columns; with this argument the computation can
be limited to a subset.}

\item{statType}{This should be either a "t","F","u","chisq" or "tlmer" depending upon the
type of statistic being thresholded.}
}
\value{
A object of type \code{mincQvals} with the same number of columns
 as the input (or the subset specified by the columns argument to
 mincFDR). Each column now contains the qvalues for each voxel. Areas
 outside the mask (if a mask was specified) will be represented by a
 value of 1. The result also has an attribute called "thresholds"
 which contains the 1, 5, 10, 15, and 20 percent false discovery rate
 thresholds.
}
\description{
Takes the output of a minc modelling function and computes False Discovery Rate thresholds.
}
\details{
This function uses the \code{qvalue} package to compute the
 False Discovery Rate threshold for the results of a \link{mincLm}
 computation. The False Discovery Rate represents the percentage of
 results expected to be a false positive. Two implementations can be
 used as specified by the method argument. "FDR" uses the
 implementation in \code{p.adjust}, whereas "pFDR" is a version of the
 postivie False Discovery Rate as found in John Storey's \code{qvalue}
 package. The main interface functions are 
 \itemize{
 \item{mincFDR.mincMultiDim}{ The workhorse function, used to compute q-values
 and thresholds for sets of minc volumes}
 \item{mincFDR.logLikRatio}{ Similar to above, but calculates thresholds by parametric
 bootstrap when possible}
 \item{mincFDR.mincSingleDim}{ Used when \link{mincLm}-like results are written out and read back in
 either to the same or another R session. In this case it loses it's \code{minMultiDim} class
 and must be converted back}
 \item{vertexFDR}{ Used with results of a \link{vertexLm}-like command. Results are converted internally
 to resemble a mincMultiDim and processed as normal}
 }
}
\section{Methods (by class)}{
\itemize{
\item \code{mincSingleDim}: mincSingleDim

\item \code{mincLogLikRatio}: mincLogLikRatio

\item \code{mincLmer}: mincLmer

\item \code{mincMultiDim}: mincMultiDim
}}
\examples{
\dontrun{
getRMINCTestData() 
# read the text file describing the dataset
gf <- read.csv("/tmp/rminctestdata/test_data_set.csv")
# run a linear model relating the data in all voxels to Genotype
vs <- mincLm(jacobians_fixed_2 ~ Sex, gf)
# compute the False Discovery Rate
qvals <- mincFDR(vs)
}
}
\seealso{
mincWriteVolume,mincLm,mincWilcoxon or mincTtest
}

