% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{mincFDRMask}
\alias{mincFDRMask}
\title{returns the specified mask, the mask associated with the buffer, or a vector of ones}
\usage{
mincFDRMask(mask, buffer)
}
\description{
returns the specified mask, the mask associated with the buffer, or a vector of ones
}

