% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_FDR.R
\name{mincFDRMask}
\alias{mincFDRMask}
\title{mincFDRMask}
\usage{
mincFDRMask(mask = NULL, buffer)
}
\arguments{
\item{mask}{a mask file or vector to be passed to mincGetMask
if left null, the buffer is checked for a mask attribute, if no mask
is found, a vector of ones is used}

\item{buffer}{a buffer describing a minc volume}
}
\value{
a numeric mask vector
}
\description{
Returns either the specified mask, the mask associated with the buffer, 
or a vector of ones to be used as a mask.
}
