% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mincGetVolume}
\alias{mincGetVolume}
\title{Return a volume as a 1D array}
\usage{
mincGetVolume(filename)
}
\arguments{
\item{filename}{A string corresponding to the location of the MINC2 file to
be read.}
}
\value{
mincLm Returns a vector of mincSingleDim class
}
\description{
Load a 3-dimensional MINC2 volume and returns it as a 1D array
}
\seealso{
mincWriteVolume
}

