% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{mincGetVolume}
\alias{mincGetVolume}
\alias{print.mincMultiDim}
\alias{print.mincSingleDim}
\title{Read a MINC file}
\usage{
mincGetVolume(filename)

print.mincMultiDim(x, ...)

print.mincSingleDim(x, ...)
}
\arguments{
\item{filename}{A string corresponding to the location of the MINC2 file to
be read.}
}
\value{
Returns a vector of mincSingleDim class
}
\description{
Load a 3-dimensional MINC2 volume and returns it as a 1D array
}
\section{Functions}{
\itemize{
\item \code{print.mincMultiDim}: print

\item \code{print.mincSingleDim}: print
}}
\examples{
\dontrun{
getRMINCTestData()
testfile <- mincGetVolume("/tmp/rminctestdata/brain_cut_out.mnc")
}
}
\seealso{
mincWriteVolume
}

