% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mincImage}
\alias{mincImage}
\title{Plot a slice from a MINC volume}
\usage{
mincImage(volume, dimension = 2, slice = NULL, low = min(volume, na.rm =
  TRUE), high = max(volume, na.rm = TRUE), reverse = FALSE,
  underTransparent = FALSE, col = gray.colors(255), scale = TRUE,
  add = FALSE, ...)
}
\arguments{
\item{volume}{a MINC volume as returned by \code{\link{mincArray}}}

\item{dimension}{the dimension (1-3) to obtain the slice from}

\item{slice}{the slice number}

\item{low}{the low end of the range to plot. If not specified then it will be
estimated based on the volume's histogram.}

\item{high}{the high end of the range to plot. If not specified then it will
be estimated based on the volume's histogram.}

\item{reverse}{whether to look only at negative numbers.}

\item{underTransparent}{whether to make anything below the low end of the
range transparent.}

\item{...}{other parameters to pass on to the \code{\link{image}} function.}
}
\description{
Calls the \code{\link{image}} plotting function from the base R graphics with
some additional data munging to make it easy to make with MINC volumes
}
\examples{
\dontrun{
mincImage(mincArray(anatVol), slice=100, col=gray.colors(255))
mincImage(mincArray(vs, 6), slice=100, col=rainbow(255),
          underTransparent = T, low=2, high=6, add=T)
}
}

