% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_lmer.R
\name{mincLmer}
\alias{mincLmer}
\title{mincified version of lmer from lme4}
\usage{
mincLmer(formula, data, mask, parallel = NULL, REML = TRUE,
  control = lmerControl(), start = NULL, verbose = 0L,
  temp_dir = getwd(), safely = FALSE, cleanup = TRUE,
  summary_type = "fixef", weights = NULL)
}
\arguments{
\item{formula}{the lmer formula, filenames go on left hand side}

\item{data}{the data frame, all items in formula should be in here}

\item{mask}{the mask within which lmer is solved}

\item{parallel}{how many processors to run on (default=single processor).
Specified as a two element vector, with the first element corresponding to
the type of parallelization, and the second to the number
of processors to use. For local running set the first element to "local" or "snowfall"
for back-compatibility, anything else will be run with batchtools see \link{pMincApply}
and \link{configureMincParallel} for details.
Leaving this argument NULL runs sequentially and may take a long time.}

\item{REML}{whether to use use Restricted Maximum Likelihood or Maximum Likelihood}

\item{control}{lmer control function}

\item{start}{lmer start function}

\item{verbose}{lmer verbosity control}

\item{temp_dir}{A directory to create temporary mask and registry files if 
\code{parallel = c("sge", n)}. This should not be \code{tempdir()} so that workers can see
these files. Defaults to the current working directory.}

\item{safely}{whether or not to wrap the per-voxel lmer code in an exception catching
block (\code{tryCatch}), when TRUE this will downgrade errors to warnings and return
NA for the result.}

\item{cleanup}{Whether or not to cleanup registry files after a queue parallelized 
run}

\item{summary_type}{Either one of
\itemize{
  \item{fixef: default and equivalent to older versions of RMINC, returns fixed effect coefficients and t-values}
  \item{ranef: returns random effect coefficients and t-values}
  \item{both: both fixed and random effects}
  \item{anova: return the F-statistic for each fixed effect}
} or a function to be used to generate the summary}

\item{weights}{weights to be applied to each observation}
}
\value{
a matrix where rows correspond to number of voxels in the file and columns to
the number of terms in the formula, with both the beta coefficient and the t-statistic
being returned. In addition, an extra column keeps the log likelihood, and another
whether the mixed effects fitting converged or not.
}
\description{
mincLmer should be used the same way as a straight lmer call, except
that the left hand side of the equation contains minc filenames rather than
an actual response variable.
}
\details{
mincLmer provides an interface to running linear mixed effects models at every
voxel. Unlike standard mincLm, however, testing hypotheses in linear mixed effects models
is more difficult, since the denominator degrees of freedom are more difficult to
determine. RMINC provides two alternatives: (1) estimating degrees of freedom using the
\code{\link{mincLmerEstimateDF}} function, and (2) comparing two separate models using
\code{\link{mincLogLikRatio}} (which in turn can be corrected using
\code{\link{mincLogLikRatioParametricBootstrap}}). For the most likely models - longitudinal
models with a separate intercept or separate intercept and slope per subject - both of these
approximations are likely correct. Be careful in using these approximations if
using more complicated random effects structures.
}
\examples{
\dontrun{
vs <- mincLmer(filenames ~ age + sex + (age|id), data=gf, mask="mask.mnc")
mincWriteVolume(vs, "age-term.mnc", "tvalue-age")
# run in parallel with multiple processors on the local machine
vs <- mincLmer(filenames ~ age + sex + (age|id), 
               data=gf, 
               mask="mask.mnc", 
               parallel=c("snowfall", 4))
# run in parallel with multiple processors over the sge batch queueing system
vs <- mincLmer(filenames ~ age + sex + (age|id), 
               data=gf, 
               mask="mask.mnc", 
               parallel=c("sge", 4))
# estimate degrees of freedom
vs <- mincLmerEstimateDF(vs)
# correct for multiple comparisons using the False Discovery Rate
(qvals <- mincFDR(vs))
# generate another model with a more complex curve for the age term
library(splines)
vs2 <- mincLmer(filenames ~ ns(age,2) + sex + (age|id), data=gf, mask="mask.mnc")
# see if that more complex age term was worth it
modelCompare <- mincLogLikRatio(vs, vs2)
mincFDR(modelCompare)
# see if there was any bias in those p-value estimates (takes a few minutes)
modelCompare <- mincLogLikRatioParametricBootstrap(modelCompare)
mincFDR(modelCompare)
}
}
\seealso{
\code{\link{lmer}} for description of lmer and lmer formulas; \code{\link{mincLm}}
}
