% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mincPairedTtest}
\alias{mincPairedTtest}
\title{Perform an paired,unequal variance t-test across a set of minc volumes}
\usage{
mincPairedTtest(filenames,grouping)
}
\arguments{
\item{filenames}{Filenames of the MINC volumes across which to run the t-test}

\item{grouping}{Contains same number of elements as
filenames; must contain exactly two groups with which to compare means. The two groups
must be the same length.}

\item{mask}{A mask specifying which voxels are to be included in the test}

\item{maskval}{The value with which to mask the data (data will masked +/- 0.5 around this value}
}
\value{
The output will be a single vector containing as many
         elements as there are voxels in the input files, with that voxel's t-statistic
}
\description{
Perform a paired,unequal variance t-test across a set of minc volumes
}
\examples{
getRMINCTestData()
gf <- read.csv("/tmp/rminctestdata/minc_summary_test_data.csv")
gf = gf[1:20,];
mptt <- mincPairedTtest(gf$jacobians_0.2,gf$Strain);
}

