% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vis2D.R
\name{mincPlotSliceSeries}
\alias{mincPlotSliceSeries}
\title{MINC Slice Series}
\usage{
mincPlotSliceSeries(anatomy, statistics = NULL, dimension = 2,
  mfrow = c(4, 5), low = NULL, high = NULL, anatLow = NULL,
  anatHigh = NULL, col = heat.colors(255), begin = 1,
  end = (dim(anatomy)[dimension] - 1), symmetric = FALSE, legend = NULL,
  locator = !is.null(legend), plottitle = NULL, indicatorLevels = NULL,
  discreteStats = FALSE)
}
\arguments{
\item{anatomy}{A minc array of the anatomy volume to plot}

\item{statistics}{optional statistics or label file to overlay on anatomy slices}

\item{dimension}{integer denoting which dimension to slice across}

\item{mfrow}{A 2 element vector of the form c(rows, columns) indicating
the number and position of slices to draw - slices are added by rows}

\item{low}{the minimum statistic to plot, calculated from the histogram if NULL}

\item{high}{the maximum statistic to plot, calculated from the histogram if NULL}

\item{anatLow}{the minimum anatomy intensity to plot, calculated from the histogram
if NULL}

\item{anatHigh}{the maximum antomy intensity to plot, calculated from the histogram
if NULL}

\item{col}{colours for statistics or for the anatomy if statistics are not passed}

\item{begin}{the first slice to plot, defaults to 1}

\item{end}{the last slice to plot, defaults to the last slice}

\item{symmetric}{whether the statistics are symmetric (such as for t-statistics)}

\item{legend}{an optional string to name the legend, indicating desire for a legend
(or not)}

\item{locator}{whether or not to draw the locator, defaults to whether or not
you requested a legend}

\item{plottitle}{the title of the plot if desired}

\item{indicatorLevels}{numeric vector indicating where to draw slice lines on the 
locator, defaults to every slice}

\item{discreteStats}{should the statistics be treated as discrete or continuous
values, useful if plotting labels}
}
\description{
Plot a simple series of slices through a minc volume
on a given dimension. Optionally include a locator
contour to show where slices are.
}
\details{
You can get a fuller tutorial on how to use the visualization tools by executing
the following command:
\code{file.show(system.file("doc/visualizationTutorial.html", package="RMINC"))}

On certain systems the slices are plotted with a reflected y-axis. To fix this
configure \code{options(RMINC_flip_image = TRUE)}
}
\examples{
\dontrun{
mincPlotSliceSeries(mincArray(anatVol),           # the anatomical volume
                    mincArray(vs, "tvalue-SexM"), # pull out one column of the stats
                    anatLow=700, anatHigh=1400,   # set anatomy thresholds
                    low=2.5, high=10,             # set stats thresholds
                    symmetric=T,                  # show separate upper and lower
                    begin=25, end=-25  ,          # remove slices from both sides  
                    legend="t-statistics")
}
}

