% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mincSelectRandomVoxels}
\alias{mincSelectRandomVoxels}
\title{selects a few random indices from a volume}
\usage{
mincSelectRandomVoxels(volumeFileName, nvoxels = 50, convert = TRUE)
}
\arguments{
\item{volumeFileName}{the filename for a MINC volume}

\item{nvoxel}{the number of voxels to select}

\item{convert}{whether to convert to MINC voxel space (default) or keep in index space}
}
\description{
Given a filename, select a few random indices using the uniform distribution
from voxels that have a value of 1 (i.e. from a mask volume)
}

