% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_voxel_statistics.R
\name{mincTFCE}
\alias{mincTFCE}
\alias{mincTFCE.mincSingleDim}
\alias{mincTFCE.matrix}
\alias{mincTFCE.mincMultiDim}
\alias{mincTFCE.mincLm}
\title{Threshold Free Cluster Enhancement}
\usage{
mincTFCE(x, ...)

\method{mincTFCE}{mincSingleDim}(x, d = 0.1, E = 0.5, H = 2,
  side = c("both", "positive", "negative"), output_file = NULL,
  keep = is.null(output_file), ...)

\method{mincTFCE}{matrix}(x, d = 0.1, E = 0.5, H = 2, side = c("both",
  "positive", "negative"), like_volume, ...)

\method{mincTFCE}{mincMultiDim}(x, d = 0.1, E = 0.5, H = 2,
  side = c("both", "positive", "negative"), like_volume = likeVolume(x),
  ...)

\method{mincTFCE}{mincLm}(x, R = 500, alternative = c("two.sided",
  "greater"), d = 0.1, E = 0.5, H = 2, side = c("both", "positive",
  "negative"), replace = FALSE, parallel = NULL, ...)
}
\arguments{
\item{x}{Either a character vector with a single filename, a \code{mincSingleDim} object,
or a \code{matrix} object, or \code{mincLm} object.}

\item{...}{additional arguments for methods}

\item{d}{The discretization step-size for approximating the threshold integral (default .1)}

\item{E}{The exponent by which to raise the extent statistic (default .5)}

\item{H}{The exponent by which to raise the height (default 2)}

\item{side}{Whether to consider positive and negative statistics or both (default both)}

\item{output_file}{A filename for the enhanced volume.}

\item{keep}{Whether or not to keep the enhanced volume, defaults to whether or not 
a \code{output_file} was specified.}

\item{like_volume}{A path to a like volume specifying the dimensions of the output volumes}

\item{R}{number of randomizations to perform}

\item{alternative}{Whether to consider a one-sided or two-sided alternative hypothesis. Default
"two-sided", use "greater" for a one sided test.}

\item{replace}{Sample with or without replacement for the randomization, defaults to FALSE (no
replacement)}

\item{parallel}{A two component vector indicating how to parallelize the computation. If the 
first element is "local" the computation will be run via the parallel package, otherwise it will
be computed using BatchJobs, see \link{pMincApply} for details. The element should be numeric
indicating the number of jobs to split the computation into.}
}
\value{
The behaviour of \code{mincTFCE} is to perform cluster free enhancement on a object,
in the single dimensional case, a string denoting a minc file or a \code{mincSingleDim} object
it returns a \code{mincSingleDim} object with the result, optionally saving the file if
keep is set to true. In the matrix case each column is converted to a \code{mincSingleDim} in
accordance with the \code{likeVolume}, this is then cluster enhanced and recomposed into a matrix.
In the mincLm case a randomization test is performed with the t-stats enhanced. The return is \itemize{
\item{TFCE: A matrix of the tvalue columns after randomization}
\item{randomization_dist: An RxT matrix where R is the number of randomizations and T is the number of t-statistic,
elements are the largest value obtained by the randomized TFCE}
\item{args: Arguments passed to the internal randomzations and TFCE code}
}
}
\description{
Perform threshold free cluster enhancement as described in 
Smith and Nichols (2008). Cluster-like structures are enhanced
to allow a hybrid cluster/voxel analysis to be performed.
}
\section{Methods (by class)}{
\itemize{
\item \code{mincSingleDim}: mincSingleDim

\item \code{matrix}: matrix

\item \code{mincMultiDim}: mincMultiDim

\item \code{mincLm}: mincLm
}}

