% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_voxel_statistics.R
\name{mincTFCE}
\alias{mincTFCE}
\title{Threshold Free Cluster Enhancement}
\usage{
mincTFCE(volume, d = 0.1, E = 0.5, H = 2, side = c("both", "positive",
  "negative"), ...)
}
\arguments{
\item{volume}{Either a character vector with a single filename, 
a \code{mincSingleDim}, or a \code{mincMultiDim} object}

\item{d}{The discretization step-size for approximating the threshold integral (default .1)}

\item{E}{The exponent by which to raise the extent statistic (default .5)}

\item{H}{The exponent by which to raise the height (default 2)}

\item{side}{Whether to consider positive and negative statistics or both (default both)}

\item{...}{additional arguments for methods}
}
\description{
Perform threshold free cluster enhancement as described in 
Smith and Nichols (2008). Cluster-like structures are enhanced
to allow a hybrid cluster/voxel analysis to be performed.
}

