% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{mincWriteVolume}
\alias{mincWriteVolume}
\alias{mincWriteVolume.mincSingleDim}
\alias{mincWriteVolume.mincMultiDim}
\alias{mincWriteVolume.default}
\title{Write a MINC volume to file}
\usage{
mincWriteVolume(buffer, ...)

\method{mincWriteVolume}{mincSingleDim}(buffer, output.filename,
  clobber = NULL, ...)

\method{mincWriteVolume}{mincMultiDim}(buffer, output.filename,
  column = 1, like.filename = NULL, clobber = NULL, ...)

\method{mincWriteVolume}{default}(buffer, output.filename, like.filename,
  clobber = NULL, ...)
}
\arguments{
\item{buffer}{The data to be written to file. Usually the result of
\link{mincLm} or some such command}

\item{...}{additional arguments to pass to methods}

\item{output.filename}{The filename to which to write the data to}

\item{clobber}{Overwrite existing output file when set to TRUE, will not 
overwrite when set to FALSE and will prompt when NULL}

\item{column}{Optional name of the column of a multidimensional MINC
object to write out. By default the first column is used}

\item{like.filename}{An existing MINC filename which has the same
dimensions as the data to be written out. Normally this information
is stored inside MINC data objects}
}
\value{
A list with the parameters of the minc volume written
}
\description{
Volume Export
}
\details{
Writes a MINC volume to file

This function takes numeric data, usually the results computed
from one of the other mincFunctions, and writes it to file so that it
can be viewed or manipulated with the standard MINC tools
}
\section{Methods (by class)}{
\itemize{
\item \code{mincSingleDim}: mincSingleDim

\item \code{mincMultiDim}: mincMultiDim

\item \code{default}: default
}}

\examples{
\dontrun{
getRMINCTestData()
# read the text file describing the dataset
gf <- read.csv("/tmp/rminctestdata/test_data_set.csv")
# run a linear model relating the data in all voxels to Sex
vs <- mincLm(gf$jacobians_fixed_2 ~ Sex, gf)
# write the results to file
mincWriteVolume(vs, "Fstat.mnc", "F-statistic")
}
}
\seealso{
mincWriteVolume,mincLm,mincFDR,mincMean,mincSd
}
