% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vis3D.R
\name{obj_montage}
\alias{obj_montage}
\title{Brain Montage}
\usage{
obj_montage(left_obj, right_obj, left_map, right_map, output = NULL,
  colour_map, colour_range = NULL, colour_default = "grey",
  colour_bar = TRUE, labels = FALSE, palette = heat.colors(255),
  symmetric = FALSE, par = list(), ..., plot_corners = c(100, 100, 900,
  900), zoom = 1, add_normals = TRUE, colour_title = "",
  close_on_output = TRUE)
}
\arguments{
\item{left_obj}{A \code{bic_obj} probably created by \link{read_obj} with the left hemisphere
of a subject's brain}

\item{right_obj}{A \code{bic_obj} probably created by \link{read_obj} with the right hemisphere
of a subject's brain}

\item{left_map}{a colour map to apply to the left hemisphere see \link{colour_mesh} for details}

\item{right_map}{a colour map to apply to the right hemisphere see \link{colour_mesh} for details}

\item{output}{Either NULL or a file path to write the snapshot.}

\item{colour_map}{either a vector with a label/measure/statistic for every vertex
or a character file path pointing to a file with such a vector in a rowwise format.}

\item{colour_range}{a two element numeric vector indicating the min and max values of 
allowable labels/measures/statistics to be includedon the surface}

\item{colour_default}{The colour given to vertices excluded by colour_range}

\item{colour_bar}{Whether or not to draw a colour bar in the figure}

\item{labels}{Whether or not the colour_map is a set of discrete labels}

\item{palette}{A palette, AKA look-up-table, providing a linear colour scale for the colours in
\code{colour_map}}

\item{symmetric}{Whether to have a positive and negative colour scale (not yet implemented)}

\item{par}{A list of plot parameters to pass to \link{add_colour_bar}}

\item{...}{additonal parameters to be passed to \link{create_mesh}}

\item{plot_corners}{The coordinates in pixels for the top left and bottom right corners of the
the rgl device. \code{c(lx, ly, rx, ry)}}

\item{zoom}{A zoom factor to apply to each subplot. This is the inverse of what you might expect
for consistency with rgl. zoom > 1 zooms out, zoom < zooms in.}

\item{add_normals}{Whether or not to add normals to the surface objects, see \link{create_mesh} for
details}

\item{colour_title}{legend title for the colour bar if requested}

\item{close_on_output}{Whether or not to close the output after taking a snapshot, defaults to
TRUE}
}
\value{
The subscenes invisibly.
}
\description{
Plot the left and right hemispheres of brain colourized by two colour maps
with a colour bar in the middle
}
\details{
This function is designed to do a simple 6 angle plot for statistic maps of the left and right
hemispheres of subject's brain. It defaults to leaving the rgl device open so that you can
take a snapshot after tweaking the angles, or changing the colour bar with \link{add_colour_bar}.
Its other mode is to take a snapshot after it has finished adding the 3d objects to the scene,
in this mode, whether or not to keep the window open can be configured with close_on_output.
}
