% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.bic_obj}
\alias{plot.bic_obj}
\title{Plot a BIC obj}
\usage{
plot.bic_obj(x, ...)
}
\arguments{
\item{x}{A \code{bic_obj} probably created by \link{read_obj}}

\item{...}{additional arguments to \link{create_mesh} including but not limited
to colour, specular, and add_normals}
}
\value{
Invisibly returns the created mesh
}
\description{
Create a basic plot of a BIC obj with a uniformly coloured mesh in the
current rgl device, opening a new one if necessary.
}

