% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vis3D.R
\name{plot.bic_obj}
\alias{plot.bic_obj}
\title{Plot a BIC obj}
\usage{
\method{plot}{bic_obj}(x, colour_map = NULL, colour_range = NULL,
  colour_default = "grey", reverse = NULL, palette = heat.colors(255),
  labels = FALSE, colour_bar = TRUE, add = FALSE, ...)
}
\arguments{
\item{x}{A \code{bic_obj} probably created by \link{read_obj}}

\item{colour_map}{A numeric vector equal in length to the number of vertices
in the \code{bic_obj} or the path to a text file with one line per vertex with
colour information.}

\item{colour_range}{a two element numeric vector indicating the min and max values of 
allowable labels/measures/statistics to be includedon the surface}

\item{colour_default}{The colour given to vertices excluded by colour_range}

\item{reverse}{Whether to have a positive and negative colour scale (not yet implemented)}

\item{palette}{A palette, AKA look-up-table, providing a linear colour scale for the colours in
\code{colour_map}}

\item{labels}{whether the statistic map should be treated as discrete labels.}

\item{colour_bar}{whether to draw a colour bar}

\item{add}{whether or not to add this object to the current rgl device (if possible)
defaults to opening a new device}

\item{...}{additional arguments to \link{create_mesh} including but not limited
to colour, specular, and add_normals}
}
\value{
invisibly returns the mesh object
}
\description{
Create a basic plot of a BIC obj in the current rgl device, opening a 
new one if necessary. If colour_map is supplied, an overlay is added to the mesh
}

