% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMINC.R
\docType{package}
\name{RMINC}
\alias{RMINC}
\alias{RMINC-package}
\title{RMINC: R interface to the MINC universe}
\description{
RMINC provides a means of accessing and performing statistical
analyses on MINC volumes. With facilities to calculate voxel-wise
statistics, perform FDR correction, rapidly visualize 
results, and much more.
}
