% Generated by roxygen2 (4.0.1): do not edit by hand
\name{runRMINCTestbed}
\alias{runRMINCTestbed}
\title{Run Testbed}
\usage{
runRMINCTestbed(verboseTest = FALSE)
}
\arguments{
\item{verboseTest}{Whether or not to verbosely print test output, default is
to print simplified results}
}
\description{
Run the test bed to ensure all RMINC functions
work on your system
}

