% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{runRMINCTestbed}
\alias{runRMINCTestbed}
\title{Run Testbed}
\usage{
runRMINCTestbed(..., verboseTest = FALSE, purgeData = TRUE)
}
\arguments{
\item{...}{additional parameter for \link[testthat]{test_dir}}

\item{verboseTest}{Whether or not to verbosely print test output, default is
to print simplified results}

\item{purgeData}{whether to remove downloaded test files
in /tmp/rminctestdata}
}
\value{
invisibly return the test results
}
\description{
Run the test bed to ensure all RMINC functions
work on your system
}

