% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_FDR.R, R/minc_voxel_statistics.R
\name{thresholds}
\alias{thresholds}
\alias{thresholds.mincQvals}
\alias{thresholds.minc_randomization}
\title{Get Probability Thresholds}
\usage{
thresholds(x, ...)

\method{thresholds}{mincQvals}(x, ...)

\method{thresholds}{minc_randomization}(x, probs = c(0.01, 0.05, 0.1, 0.2), ...)
}
\arguments{
\item{x}{A \code{mincQvals} object, typically computed with \code{mincFDR} or a 
\code{minc*_randomzation} type object.
methods}

\item{...}{extra arguments for methods}

\item{probs}{What probabilities to compute thresholds for (only applicable with randomization objects)}
}
\value{
A matrix of thresholds, accessible with standard matrix indexing
}
\description{
Get Probability Thresholds
}
\section{Methods (by class)}{
\itemize{
\item \code{mincQvals}: mincQvals

\item \code{minc_randomization}: minc_randomization
}}

