% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vertex_statistics.R
\name{vertexAnova}
\alias{vertexAnova}
\title{Performs ANOVA on each vertex point specified}
\usage{
vertexAnova(formula, data, subset = NULL, column = 1)
}
\arguments{
\item{formula}{a model formula}

\item{data}{a data.frame containing variables in formula}

\item{subset}{rows to be used, by default all are used}

\item{column}{Which column to treat as the input from vertex files.}
}
\value{
Returns an array with the F-statistic for each model specified by formula with the following attributes: 
\itemize{
\item{model}{ design matrix}
\item{filenames}{ minc file names input}
\item{dimensions}{ dimensions of the statistics matrix}
\item{dimnames}{ names of the dimensions for the statistic matrix}
\item{stat-type}{ types of statistic used}
\item{df}{ degrees of freedom of each statistic}
}
}
\description{
Performs ANOVA on each vertex point specified
}
\examples{
\dontrun{
getRMINCTestData() 
gf = read.csv("/tmp/rminctestdata/CIVET_TEST.csv")
gf = civet.getAllFilenames(gf,"ID","TEST","/tmp/rminctestdata/CIVET","TRUE","1.1.12")
gf = civet.readAllCivetFiles("/tmp/rminctestdata/AAL.csv",gf)
result = vertexAnova(CIVETFILES$nativeRMStlink20mmleft~Primary.Diagnosis,gf)
}
}
\seealso{
mincAnova,anatAnova
}
