% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vertex_statistics.R
\name{vertexApply}
\alias{vertexApply}
\title{Apply function over vertex Files}
\usage{
vertexApply(
  filenames,
  fun,
  ...,
  mask = NULL,
  parallel = NULL,
  collate = simplify_masked,
  transpose = FALSE,
  column = 1
)
}
\arguments{
\item{filenames}{vertex file names}

\item{fun}{A function to be applied to each vertex}

\item{...}{additional arguments to \code{fun}}

\item{mask}{A vector of filename indicating a vertex mask (\code{fun} is applied to all vertices
where mask is greater than .5)}

\item{parallel}{A two component vector indicating how to parallelize the computation. If the 
first element is "local" the computation will be run via the parallel package, otherwise it will
be computed using batchtools, see \link{pMincApply} for details. The element should be numeric
indicating the number of jobs to split the computation into.}

\item{collate}{A function to reduce the (potentially masked) list of results into a nice
structure. Defaults to \link{simplify_masked}}

\item{transpose}{Whether to alternatively transpose the vertex matrix and apply a function to
each subject}

\item{column}{Which column to treat as the input from vertex files.}
}
\value{
The a matrix with a row of results for each vertex
}
\description{
This function is used to compute an arbitrary function of every region in a set of vertex files.
}
\examples{
\dontrun{
getRMINCTestData() 
gf = read.csv("/tmp/rminctestdata/CIVET_TEST.csv")
gf = civet.getAllFilenames(gf,"ID","TEST","/tmp/rminctestdata/CIVET","TRUE","1.1.12")
gf = civet.readAllCivetFiles("/tmp/rminctestdata/AAL.csv",gf)
vm <- vertexApply(gf$CIVETFILES$nativeRMStlink20mmleft, mean)
}
}
