% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_interface.R
\name{vertexLm}
\alias{vertexLm}
\title{Calculates statistics and coefficients for linear model of specified vertex files}
\usage{
vertexLm(formula, data, subset = NULL)
}
\arguments{
\item{formula}{a model formula. The RHS of the formula may contain one term with filenames. If
so only the + operator may be used, and only two terms may appear on the RHS}

\item{data}{a data.frame containing variables in formula}

\item{subset}{rows to be used, by default all are used}
}
\value{
Returns an object containing the R-Squared value,beta coefficients, F 
and t statistcs that can be passed directly into vertexFDR.
}
\description{
Calculates statistics and coefficients for linear model of specified vertex files
}
\examples{
\dontrun{ 
getRMINCTestData() 
gf = read.csv("/tmp/rminctestdata/CIVET_TEST.csv")
gf = civet.getAllFilenames(gf,"ID","TEST","/tmp/rminctestdata/CIVET","TRUE","1.1.12")
gf = civet.readAllCivetFiles("/tmp/rminctestdata/AAL.csv",gf)
result = vertexLm(CIVETFILES$nativeRMStlink20mmleft~Primary.Diagnosis,gf) 
}
}
\seealso{
mincLm,anatLm,vertexFDR
}

