% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vertex_statistics.R
\name{vertexLmer}
\alias{vertexLmer}
\title{Vertex Mixed Effects Models}
\usage{
vertexLmer(formula, data, mask = NULL, parallel = NULL, REML = TRUE,
  control = lmerControl(), start = NULL, verbose = 0L, safely = FALSE,
  summary_type = "fixef")
}
\arguments{
\item{formula}{the lmer formula, filenames go on left hand side}

\item{data}{the data frame, all items in formula should be in here}

\item{mask}{the mask within which lmer is solved}

\item{parallel}{how many processors to run on (default=single processor).
Specified as a two element vector, with the first element corresponding to
the type of parallelization, and the second to the number
of processors to use. For local running set the first element to "local" or "snowfall"
for back-compatibility, anything else will be run with batchtools see \link{pMincApply}.
Leaving this argument NULL runs sequentially and may take a long time.}

\item{REML}{whether to use use Restricted Maximum Likelihood or Maximum Likelihood}

\item{control}{lmer control function}

\item{start}{lmer start function}

\item{verbose}{lmer verbosity control}

\item{safely}{whether or not to wrap the per-voxel lmer code in an exception catching
block (\code{tryCatch}), when TRUE this will downgrade errors to warnings and return
NA for the result.}

\item{summary_type}{Either one of
\itemize{
  \item{fixef: default and equivalent to older versions of RMINC, returns fixed effect coefficients and t-values}
  \item{ranef: returns random effect coefficients and t-values}
  \item{both: both fixed and random effects}
  \item{anova: return the F-statistic for each fixed effect}
} or a function to be used to generate the summary}
}
\description{
Perform linear mixed effects model fitting for vertex data.
vertexLmer should be used the same way as a straight lmer call, except
that the left hand side of the equation contains vertex filenames rather than
an actual response variable.
}
\details{
\code{vertexLmer}, like its relative \link{mincLmer} provides an interface to running 
linear mixed effects models at every vertex. Unlike standard linear models testing hypotheses 
in linear mixed effects models is more difficult, since the denominator degrees of freedom are 
more difficult to  determine. RMINC provides estimating degrees of freedom using the
\code{\link{vertexLmerEstimateDF}} function. For the most likely models - longitudinal
models with a separate intercept or separate intercept and slope per subject - this
approximation is likely correct. Be careful in using this approximation if
using more complicated random effects structures.
}
\seealso{
\code{\link{lmer}} for description of lmer and lmer formulas; \code{\link{mincLm}}
}
