% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vertex_statistics.R
\name{vertexSummaries}
\alias{vertexSummaries}
\alias{vertexMean}
\alias{vertexSum}
\alias{vertexVar}
\alias{vertexSd}
\title{Create descriptive statistics across a series of vertex files}
\usage{
vertexMean(filenames, column = 1)

vertexSum(filenames, column = 1)

vertexVar(filenames, column = 1)

vertexSd(filenames, column = 1)
}
\arguments{
\item{filenames}{Filenames of the vertex volumes across which to create the
descriptive statistic.}

\item{column}{Which column to treat as the input from vertex files.}
}
\value{
\item{out}{The output will be a single vector containing as many
elements as there are vertices in the input files.}
}
\description{
This function is used to compute the mean, standard deviation, sum, or
variance of every vertex in a set of vertex files.
}
\section{Functions}{
\itemize{
\item \code{vertexMean()}: mean

\item \code{vertexSum()}: sum

\item \code{vertexVar()}: var

\item \code{vertexSd()}: standard deviation

}}
\examples{

\dontrun{
# read the text file describing the dataset
gf <- read.csv("control-file.csv")
# compute the mean at every voxel of all files.
means <- vertexMean(gf$filenames)
}
}
\seealso{
vertexLm
}
