% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_vertex_statistics.R
\name{vertexTFCE}
\alias{vertexTFCE}
\alias{vertexTFCE.numeric}
\title{Threshold Free Cluster Enhancement}
\usage{
vertexTFCE(x, ...)

\method{vertexTFCE}{numeric}(x, surface, d = 0.1, E = 0.5, H = 2,
  nsteps = 100, side = c("both", "positive", "negative"), weights = NULL,
  ...)
}
\arguments{
\item{x}{A numeric vector, a filepath to a set of values,
or a \code{matrix} object, or \code{vertexLm} object.}

\item{...}{additional arguments for methods}

\item{surface}{Either a mesh object corresponding to the surface, or
an igraph graph object representing the surface.
\link{obj_to_graph}}

\item{d}{The discretization step-size for approximating the threshold integral (default .1)}

\item{E}{The exponent by which to raise the extent statistic (default .5)}

\item{H}{The exponent by which to raise the height (default 2)}

\item{nsteps}{The number of steps to discrete}

\item{side}{Whether to consider positive and negative statistics or both (default both)}

\item{weights}{A weighting vector assigning area to vertices, by default area is computed as
the 1/3 the sum of the areas of each triangle the vertex is a member of. Another reasonable option
is 1, corresponding to an unweighted TFCE.}
}
\value{
The behaviour of \code{vertexTFCE} is to perform cluster free enhancement on a object,
in the single dimensional case, a string denoting a vertex file or a numeric vector
it returns a numeric vector with the result. In the matrix case each column is cluster enhanced and 
recomposed into a matrix.
In the \link{vertexLm} case a randomization test is performed with the t-stats enhanced computing the maximal
value after enhancement. A vector of maxima is returned to compute exceedence probabilities, along with
the original statistic after cluster enhancement.
}
\description{
Perform threshold free cluster enhancement as described in 
Smith and Nichols (2008). Cluster-like structures are enhanced
to allow a hybrid cluster/voxel analysis to be performed.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: numeric
}}

