% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minc_misc.R
\name{volume.combineMaskVolumes}
\alias{volume.combineMaskVolumes}
\title{Combine Multiple Mask Volumes into a Single Mask Volume}
\usage{
volume.combineMaskVolumes(vol_list)
}
\arguments{
\item{vol_list}{A list containing more than one mask volume.  Note that all 
volumes must reflect the same sampling.}
}
\value{
A single aggregate MincVolumeIO volume is returned.
}
\description{
Given a list containing more than one label volume, combine those 
volumes, creating an aggregate mask volume.
}
\author{
Jim Nikelski \email{nikelski@bic.mni.mcgill.ca}
}

